Add documentation directory
