--- baseProject/temp/jsp/org/apache/jsp/pages/getzipcodes_jsp.java
@@ -1,110 +1,0 @@
-/*
- * Generated by the Jasper component of Apache Tomcat
- * Version: jetty/9.2.3.v20140905
- * Generated at: 2016-10-04 18:23:05 UTC
- * Note: The last modified time of this file was set to
- *       the last modified time of the source file after
- *       generation to assist with modification tracking.
- */
-package org.apache.jsp.pages;
-
-import javax.servlet.*;
-import javax.servlet.http.*;
-import javax.servlet.jsp.*;
-import java.util.Iterator;
-import ch.unibe.ese.team1.model.Location;
-import java.util.List;
-import ch.unibe.ese.team1.controller.service.GeoDataService;
-import org.springframework.web.servlet.support.RequestContextUtils;
-import org.springframework.context.ApplicationContext;
-
-public final class getzipcodes_jsp extends org.apache.jasper.runtime.HttpJspBase
-    implements org.apache.jasper.runtime.JspSourceDependent {
-
-  private static final javax.servlet.jsp.JspFactory _jspxFactory =
-          javax.servlet.jsp.JspFactory.getDefaultFactory();
-
-  private static java.util.Map<java.lang.String,java.lang.Long> _jspx_dependants;
-
-  private javax.el.ExpressionFactory _el_expressionfactory;
-  private org.apache.tomcat.InstanceManager _jsp_instancemanager;
-
-  public java.util.Map<java.lang.String,java.lang.Long> getDependants() {
-    return _jspx_dependants;
-  }
-
-  public void _jspInit() {
-    _el_expressionfactory = _jspxFactory.getJspApplicationContext(getServletConfig().getServletContext()).getExpressionFactory();
-    _jsp_instancemanager = org.apache.jasper.runtime.InstanceManagerFactory.getInstanceManager(getServletConfig());
-  }
-
-  public void _jspDestroy() {
-  }
-
-  public void _jspService(final javax.servlet.http.HttpServletRequest request, final javax.servlet.http.HttpServletResponse response)
-        throws java.io.IOException, javax.servlet.ServletException {
-
-    final javax.servlet.jsp.PageContext pageContext;
-    javax.servlet.http.HttpSession session = null;
-    final javax.servlet.ServletContext application;
-    final javax.servlet.ServletConfig config;
-    javax.servlet.jsp.JspWriter out = null;
-    final java.lang.Object page = this;
-    javax.servlet.jsp.JspWriter _jspx_out = null;
-    javax.servlet.jsp.PageContext _jspx_page_context = null;
-
-
-    try {
-      response.setContentType("text/html; charset=ISO-8859-1");
-      pageContext = _jspxFactory.getPageContext(this, request, response,
-      			null, true, 8192, true);
-      _jspx_page_context = pageContext;
-      application = pageContext.getServletContext();
-      config = pageContext.getServletConfig();
-      session = pageContext.getSession();
-      out = pageContext.getOut();
-      _jspx_out = out;
-
-      out.write("\r\n");
-      out.write("\r\n");
-      out.write("\r\n");
-      out.write("\r\n");
-      out.write("\r\n");
-      out.write("\r\n");
-      out.write("\r\n");
-
-	ApplicationContext ac = RequestContextUtils
-	.getWebApplicationContext(request);
-	GeoDataService geoDataService = (GeoDataService) ac
-	.getBean(GeoDataService.class);
-	
-	List<Location> zipCodes = geoDataService.getAllLocations();
-	Iterator<Location> iterator = zipCodes.iterator();
-	out.print("[");
-	Location zip = iterator.next();
-	out.print("\"" + zip.getZip() + " - " + zip.getCity() + "\"");
-	while(iterator.hasNext()){
-		zip = iterator.next();
-		out.print(",\"" + zip.getZip() + " - " + zip.getCity() + "\"");
-	}
-	out.print("]");
-
-    } catch (java.lang.Throwable t) {
-      if (!(t instanceof javax.servlet.jsp.SkipPageException)){
-        out = _jspx_out;
-        if (out != null && out.getBufferSize() != 0)
-          try {
-            if (response.isCommitted()) {
-              out.flush();
-            } else {
-              out.clearBuffer();
-            }
-          } catch (java.io.IOException e) {}
-        if (_jspx_page_context != null) _jspx_page_context.handlePageException(t);
-        else throw new ServletException(t);
-      }
-    } finally {
-      _jspxFactory.releasePageContext(_jspx_page_context);
-    }
-  }
-}
