--- baseProject/temp/jsp/org/apache/jsp/pages/results_jsp.java
@@ -1,1681 +1,0 @@
-/*
- * Generated by the Jasper component of Apache Tomcat
- * Version: jetty/9.2.3.v20140905
- * Generated at: 2016-10-04 18:23:16 UTC
- * Note: The last modified time of this file was set to
- *       the last modified time of the source file after
- *       generation to assist with modification tracking.
- */
-package org.apache.jsp.pages;
-
-import javax.servlet.*;
-import javax.servlet.http.*;
-import javax.servlet.jsp.*;
-
-public final class results_jsp extends org.apache.jasper.runtime.HttpJspBase
-    implements org.apache.jasper.runtime.JspSourceDependent {
-
-  private static final javax.servlet.jsp.JspFactory _jspxFactory =
-          javax.servlet.jsp.JspFactory.getDefaultFactory();
-
-  private static java.util.Map<java.lang.String,java.lang.Long> _jspx_dependants;
-
-  static {
-    _jspx_dependants = new java.util.HashMap<java.lang.String,java.lang.Long>(7);
-    _jspx_dependants.put("file:/C:/Users/Julian%20Schuhmacher/.m2/repository/org/springframework/spring-webmvc/4.1.1.RELEASE/spring-webmvc-4.1.1.RELEASE.jar", Long.valueOf(1475180010191L));
-    _jspx_dependants.put("jar:file:/C:/Users/Julian%20Schuhmacher/.m2/repository/org/apache/taglibs/taglibs-standard-impl/1.2.1/taglibs-standard-impl-1.2.1.jar!/META-INF/c.tld", Long.valueOf(1384364862000L));
-    _jspx_dependants.put("jar:file:/C:/Users/Julian%20Schuhmacher/.m2/repository/javax/servlet/jstl/1.2/jstl-1.2.jar!/META-INF/fmt.tld", Long.valueOf(1153377882000L));
-    _jspx_dependants.put("file:/C:/Users/Julian%20Schuhmacher/.m2/repository/org/apache/taglibs/taglibs-standard-impl/1.2.1/taglibs-standard-impl-1.2.1.jar", Long.valueOf(1475217457864L));
-    _jspx_dependants.put("jar:file:/C:/Users/Julian%20Schuhmacher/.m2/repository/org/springframework/spring-webmvc/4.1.1.RELEASE/spring-webmvc-4.1.1.RELEASE.jar!/META-INF/spring-form.tld", Long.valueOf(1412119640000L));
-    _jspx_dependants.put("jar:file:/C:/Users/Julian%20Schuhmacher/.m2/repository/org/springframework/spring-webmvc/4.1.1.RELEASE/spring-webmvc-4.1.1.RELEASE.jar!/META-INF/spring.tld", Long.valueOf(1412119640000L));
-    _jspx_dependants.put("file:/C:/Users/Julian%20Schuhmacher/.m2/repository/javax/servlet/jstl/1.2/jstl-1.2.jar", Long.valueOf(1475180013346L));
-  }
-
-  private org.apache.jasper.runtime.TagHandlerPool _005fjspx_005ftagPool_005fc_005fimport_0026_005furl_005fnobody;
-  private org.apache.jasper.runtime.TagHandlerPool _005fjspx_005ftagPool_005fc_005fchoose;
-  private org.apache.jasper.runtime.TagHandlerPool _005fjspx_005ftagPool_005fc_005fwhen_0026_005ftest;
-  private org.apache.jasper.runtime.TagHandlerPool _005fjspx_005ftagPool_005fc_005fotherwise;
-  private org.apache.jasper.runtime.TagHandlerPool _005fjspx_005ftagPool_005fc_005fforEach_0026_005fvar_005fitems;
-  private org.apache.jasper.runtime.TagHandlerPool _005fjspx_005ftagPool_005fc_005furl_0026_005fvalue_005fnobody;
-  private org.apache.jasper.runtime.TagHandlerPool _005fjspx_005ftagPool_005ffmt_005fformatDate_0026_005fvar_005fvalue_005ftype_005fpattern_005fnobody;
-  private org.apache.jasper.runtime.TagHandlerPool _005fjspx_005ftagPool_005fform_005fform_0026_005fmodelAttribute_005fmethod_005fid_005fautocomplete_005faction;
-  private org.apache.jasper.runtime.TagHandlerPool _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fpath_005fname_005fid_005fnobody;
-  private org.apache.jasper.runtime.TagHandlerPool _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fstyle_005fpath_005fname_005fid_005fnobody;
-  private org.apache.jasper.runtime.TagHandlerPool _005fjspx_005ftagPool_005fform_005ferrors_0026_005fpath_005fcssClass_005fnobody;
-  private org.apache.jasper.runtime.TagHandlerPool _005fjspx_005ftagPool_005fform_005finput_0026_005ftype_005ftabindex_005fplaceholder_005fpath_005fname_005fid_005fnobody;
-  private org.apache.jasper.runtime.TagHandlerPool _005fjspx_005ftagPool_005fform_005finput_0026_005ftype_005fstep_005fplaceholder_005fpath_005fid_005fnobody;
-  private org.apache.jasper.runtime.TagHandlerPool _005fjspx_005ftagPool_005fform_005finput_0026_005ftype_005fpath_005fid_005fnobody;
-  private org.apache.jasper.runtime.TagHandlerPool _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fvalue_005fpath_005fid_005fnobody;
-
-  private javax.el.ExpressionFactory _el_expressionfactory;
-  private org.apache.tomcat.InstanceManager _jsp_instancemanager;
-
-  public java.util.Map<java.lang.String,java.lang.Long> getDependants() {
-    return _jspx_dependants;
-  }
-
-  public void _jspInit() {
-    _005fjspx_005ftagPool_005fc_005fimport_0026_005furl_005fnobody = org.apache.jasper.runtime.TagHandlerPool.getTagHandlerPool(getServletConfig());
-    _005fjspx_005ftagPool_005fc_005fchoose = org.apache.jasper.runtime.TagHandlerPool.getTagHandlerPool(getServletConfig());
-    _005fjspx_005ftagPool_005fc_005fwhen_0026_005ftest = org.apache.jasper.runtime.TagHandlerPool.getTagHandlerPool(getServletConfig());
-    _005fjspx_005ftagPool_005fc_005fotherwise = org.apache.jasper.runtime.TagHandlerPool.getTagHandlerPool(getServletConfig());
-    _005fjspx_005ftagPool_005fc_005fforEach_0026_005fvar_005fitems = org.apache.jasper.runtime.TagHandlerPool.getTagHandlerPool(getServletConfig());
-    _005fjspx_005ftagPool_005fc_005furl_0026_005fvalue_005fnobody = org.apache.jasper.runtime.TagHandlerPool.getTagHandlerPool(getServletConfig());
-    _005fjspx_005ftagPool_005ffmt_005fformatDate_0026_005fvar_005fvalue_005ftype_005fpattern_005fnobody = org.apache.jasper.runtime.TagHandlerPool.getTagHandlerPool(getServletConfig());
-    _005fjspx_005ftagPool_005fform_005fform_0026_005fmodelAttribute_005fmethod_005fid_005fautocomplete_005faction = org.apache.jasper.runtime.TagHandlerPool.getTagHandlerPool(getServletConfig());
-    _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fpath_005fname_005fid_005fnobody = org.apache.jasper.runtime.TagHandlerPool.getTagHandlerPool(getServletConfig());
-    _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fstyle_005fpath_005fname_005fid_005fnobody = org.apache.jasper.runtime.TagHandlerPool.getTagHandlerPool(getServletConfig());
-    _005fjspx_005ftagPool_005fform_005ferrors_0026_005fpath_005fcssClass_005fnobody = org.apache.jasper.runtime.TagHandlerPool.getTagHandlerPool(getServletConfig());
-    _005fjspx_005ftagPool_005fform_005finput_0026_005ftype_005ftabindex_005fplaceholder_005fpath_005fname_005fid_005fnobody = org.apache.jasper.runtime.TagHandlerPool.getTagHandlerPool(getServletConfig());
-    _005fjspx_005ftagPool_005fform_005finput_0026_005ftype_005fstep_005fplaceholder_005fpath_005fid_005fnobody = org.apache.jasper.runtime.TagHandlerPool.getTagHandlerPool(getServletConfig());
-    _005fjspx_005ftagPool_005fform_005finput_0026_005ftype_005fpath_005fid_005fnobody = org.apache.jasper.runtime.TagHandlerPool.getTagHandlerPool(getServletConfig());
-    _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fvalue_005fpath_005fid_005fnobody = org.apache.jasper.runtime.TagHandlerPool.getTagHandlerPool(getServletConfig());
-    _el_expressionfactory = _jspxFactory.getJspApplicationContext(getServletConfig().getServletContext()).getExpressionFactory();
-    _jsp_instancemanager = org.apache.jasper.runtime.InstanceManagerFactory.getInstanceManager(getServletConfig());
-  }
-
-  public void _jspDestroy() {
-    _005fjspx_005ftagPool_005fc_005fimport_0026_005furl_005fnobody.release();
-    _005fjspx_005ftagPool_005fc_005fchoose.release();
-    _005fjspx_005ftagPool_005fc_005fwhen_0026_005ftest.release();
-    _005fjspx_005ftagPool_005fc_005fotherwise.release();
-    _005fjspx_005ftagPool_005fc_005fforEach_0026_005fvar_005fitems.release();
-    _005fjspx_005ftagPool_005fc_005furl_0026_005fvalue_005fnobody.release();
-    _005fjspx_005ftagPool_005ffmt_005fformatDate_0026_005fvar_005fvalue_005ftype_005fpattern_005fnobody.release();
-    _005fjspx_005ftagPool_005fform_005fform_0026_005fmodelAttribute_005fmethod_005fid_005fautocomplete_005faction.release();
-    _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fpath_005fname_005fid_005fnobody.release();
-    _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fstyle_005fpath_005fname_005fid_005fnobody.release();
-    _005fjspx_005ftagPool_005fform_005ferrors_0026_005fpath_005fcssClass_005fnobody.release();
-    _005fjspx_005ftagPool_005fform_005finput_0026_005ftype_005ftabindex_005fplaceholder_005fpath_005fname_005fid_005fnobody.release();
-    _005fjspx_005ftagPool_005fform_005finput_0026_005ftype_005fstep_005fplaceholder_005fpath_005fid_005fnobody.release();
-    _005fjspx_005ftagPool_005fform_005finput_0026_005ftype_005fpath_005fid_005fnobody.release();
-    _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fvalue_005fpath_005fid_005fnobody.release();
-  }
-
-  public void _jspService(final javax.servlet.http.HttpServletRequest request, final javax.servlet.http.HttpServletResponse response)
-        throws java.io.IOException, javax.servlet.ServletException {
-
-    final javax.servlet.jsp.PageContext pageContext;
-    javax.servlet.http.HttpSession session = null;
-    final javax.servlet.ServletContext application;
-    final javax.servlet.ServletConfig config;
-    javax.servlet.jsp.JspWriter out = null;
-    final java.lang.Object page = this;
-    javax.servlet.jsp.JspWriter _jspx_out = null;
-    javax.servlet.jsp.PageContext _jspx_page_context = null;
-
-
-    try {
-      response.setContentType("text/html;charset=utf-8");
-      pageContext = _jspxFactory.getPageContext(this, request, response,
-      			null, true, 8192, true);
-      _jspx_page_context = pageContext;
-      application = pageContext.getServletContext();
-      config = pageContext.getServletConfig();
-      session = pageContext.getSession();
-      out = pageContext.getOut();
-      _jspx_out = out;
-
-      out.write("\r\n");
-      out.write("\r\n");
-      out.write("\r\n");
-      out.write("\r\n");
-      out.write("\r\n");
-      out.write("\r\n");
-      out.write("\r\n");
-      if (_jspx_meth_c_005fimport_005f0(_jspx_page_context))
-        return;
-      out.write("\r\n");
-      out.write("<pre><a href=\"/\">Home</a>   &gt;   <a href=\"/searchAd/\">Search</a>   &gt;   Results</pre>\r\n");
-      out.write("\r\n");
-      out.write("<script>\r\n");
-      out.write("function validateType(form)\r\n");
-      out.write("{\r\n");
-      out.write("\tvar room = document.getElementById('room');\r\n");
-      out.write("\tvar studio = document.getElementById('studio');\r\n");
-      out.write("\tvar neither = document.getElementById('neither');\r\n");
-      out.write("\tvar both = document.getElementById('both');\r\n");
-      out.write("\tvar type = document.getElementById('type');\r\n");
-      out.write("\tvar filtered = document.getElementById('filtered');\r\n");
-      out.write("\t\r\n");
-      out.write("\tif(room.checked && studio.checked) {\r\n");
-      out.write("\t\tboth.checked = true;\r\n");
-      out.write("\t\tneither.checked = false;\r\n");
-      out.write("\t}\r\n");
-      out.write("\telse if(!room.checked && !studio.checked) {\r\n");
-      out.write("\t\tboth.checked = false;\r\n");
-      out.write("\t\tneither.checked = true;\r\n");
-      out.write("\t}\r\n");
-      out.write("\telse {\r\n");
-      out.write("\t\tboth.checked = false;\r\n");
-      out.write("\t\tneither.checked = false;\r\n");
-      out.write("\t\ttype.checked = studio.checked;\r\n");
-      out.write("\t}\r\n");
-      out.write("\tfiltered.checked = true;\r\n");
-      out.write("}\r\n");
-      out.write("</script>\r\n");
-      out.write("\r\n");
-      out.write("<script>\r\n");
-      out.write("/*\r\n");
-      out.write(" * This script takes all the resultAd divs and sorts them by a parameter specified by the user.\r\n");
-      out.write(" * No arguments need to be passed, since the function simply looks up the dropdown selection.\r\n");
-      out.write(" */\r\n");
-      out.write("function sort_div_attribute() {\r\n");
-      out.write("    //determine sort modus (by which attribute, asc/desc)\r\n");
-      out.write("    var sortmode = $('#modus').find(\":selected\").val();   \r\n");
-      out.write("    \r\n");
-      out.write("    //only start the process if a modus has been selected\r\n");
-      out.write("    if(sortmode.length > 0) {\r\n");
-      out.write("    \tvar attname;\r\n");
-      out.write("\t\t\r\n");
-      out.write("    \t//determine which variable we pass to the sort function\r\n");
-      out.write("\t\tif(sortmode == \"price_asc\" || sortmode == \"price_desc\")\r\n");
-      out.write("\t\t\tattname = 'data-price';\r\n");
-      out.write("\t    else if(sortmode == \"moveIn_asc\" || sortmode == \"moveIn_desc\")\t\r\n");
-      out.write("\t\t\tattname = 'data-moveIn';\r\n");
-      out.write("\t    else\r\n");
-      out.write("\t\t\tattname = 'data-age';\r\n");
-      out.write("    \t\r\n");
-      out.write("\t\t//copying divs into an array which we're going to sort\r\n");
-      out.write("\t    var divsbucket = new Array();\r\n");
-      out.write("\t    var divslist = $('div.resultAd');\r\n");
-      out.write("\t    var divlength = divslist.length;\r\n");
-      out.write("\t    for (a = 0; a < divlength; a++) {\r\n");
-      out.write("\t\t\tdivsbucket[a] = new Array();\r\n");
-      out.write("\t\t\tdivsbucket[a][0] = divslist[a].getAttribute(attname);\r\n");
-      out.write("\t\t\tdivsbucket[a][1] = divslist[a];\r\n");
-      out.write("\t\t\tdivslist[a].remove();\r\n");
-      out.write("\t    }\r\n");
-      out.write("\t\t\r\n");
-      out.write("\t    //sort the array\r\n");
-      out.write("\t\tdivsbucket.sort(function(a, b) {\r\n");
-      out.write("\t    if (a[0] == b[0])\r\n");
-      out.write("\t\t\treturn 0;\r\n");
-      out.write("\t    else if (a[0] > b[0])\r\n");
-      out.write("\t\t\treturn 1;\r\n");
-      out.write("        else\r\n");
-      out.write("\t\t\treturn -1;\r\n");
-      out.write("\t\t});\r\n");
-      out.write("\r\n");
-      out.write("\t    //invert sorted array for certain sort options\r\n");
-      out.write("\t\tif(sortmode == \"price_desc\" || sortmode == \"moveIn_asc\" || sortmode == \"dateAge_asc\")\r\n");
-      out.write("\t\t\tdivsbucket.reverse();\r\n");
-      out.write("        \r\n");
-      out.write("\t    //insert sorted divs into document again\r\n");
-      out.write("\t\tfor(a = 0; a < divlength; a++)\r\n");
-      out.write("        \t$(\"#resultsDiv\").append($(divsbucket[a][1]));\r\n");
-      out.write("\t}\r\n");
-      out.write("}\r\n");
-      out.write("</script>\r\n");
-      out.write("\r\n");
-      out.write("<script>\r\n");
-      out.write("\t$(document).ready(function() {\r\n");
-      out.write("\t\t$(\"#city\").autocomplete({\r\n");
-      out.write("\t\t\tminLength : 2\r\n");
-      out.write("\t\t});\r\n");
-      out.write("\t\t$(\"#city\").autocomplete({\r\n");
-      out.write("\t\t\tsource : ");
-      if (_jspx_meth_c_005fimport_005f1(_jspx_page_context))
-        return;
-      out.write("\r\n");
-      out.write("\t\t});\r\n");
-      out.write("\t\t$(\"#city\").autocomplete(\"option\", {\r\n");
-      out.write("\t\t\tenabled : true,\r\n");
-      out.write("\t\t\tautoFocus : true\r\n");
-      out.write("\t\t});\r\n");
-      out.write("\t\t\r\n");
-      out.write("\t\t$(\"#field-earliestMoveInDate\").datepicker({\r\n");
-      out.write("\t\t\tdateFormat : 'dd-mm-yy'\r\n");
-      out.write("\t\t});\r\n");
-      out.write("\t\t$(\"#field-latestMoveInDate\").datepicker({\r\n");
-      out.write("\t\t\tdateFormat : 'dd-mm-yy'\r\n");
-      out.write("\t\t});\r\n");
-      out.write("\t\t$(\"#field-earliestMoveOutDate\").datepicker({\r\n");
-      out.write("\t\t\tdateFormat : 'dd-mm-yy'\r\n");
-      out.write("\t\t});\r\n");
-      out.write("\t\t$(\"#field-latestMoveOutDate\").datepicker({\r\n");
-      out.write("\t\t\tdateFormat : 'dd-mm-yy'\r\n");
-      out.write("\t\t});\r\n");
-      out.write("\t});\r\n");
-      out.write("</script>\r\n");
-      out.write("\r\n");
-      out.write("<h1>Search results:</h1>\r\n");
-      out.write("\r\n");
-      out.write("<hr />\r\n");
-      out.write("\r\n");
-      out.write("<div>\r\n");
-      out.write("<select id=\"modus\">\r\n");
-      out.write("    <option value=\"\">Sort by:</option>\r\n");
-      out.write("    <option value=\"price_asc\">Price (ascending)</option>\r\n");
-      out.write("    <option value=\"price_desc\">Price (descending)</option>\r\n");
-      out.write("    <option value=\"moveIn_desc\">Move-in date (earliest to latest)</option>\r\n");
-      out.write("    <option value=\"moveIn_asc\">Move-in date (latest to earliest)</option>\r\n");
-      out.write("    <option value=\"dateAge_asc\">Date created (youngest to oldest)</option>\r\n");
-      out.write("    <option value=\"dateAge_desc\">Date created (oldest to youngest)</option>\r\n");
-      out.write("</select>\r\n");
-      out.write("\r\n");
-      out.write("<button onClick=\"sort_div_attribute()\">Sort</button>\t\r\n");
-      out.write("</div>\r\n");
-      if (_jspx_meth_c_005fchoose_005f0(_jspx_page_context))
-        return;
-      out.write("\r\n");
-      out.write("\r\n");
-      //  form:form
-      org.springframework.web.servlet.tags.form.FormTag _jspx_th_form_005fform_005f0 = (org.springframework.web.servlet.tags.form.FormTag) _005fjspx_005ftagPool_005fform_005fform_0026_005fmodelAttribute_005fmethod_005fid_005fautocomplete_005faction.get(org.springframework.web.servlet.tags.form.FormTag.class);
-      _jspx_th_form_005fform_005f0.setPageContext(_jspx_page_context);
-      _jspx_th_form_005fform_005f0.setParent(null);
-      // /pages/results.jsp(176,0) name = method type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-      _jspx_th_form_005fform_005f0.setMethod("post");
-      // /pages/results.jsp(176,0) name = modelAttribute type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-      _jspx_th_form_005fform_005f0.setModelAttribute("searchForm");
-      // /pages/results.jsp(176,0) name = action type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-      _jspx_th_form_005fform_005f0.setAction("/results");
-      // /pages/results.jsp(176,0) name = id type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-      _jspx_th_form_005fform_005f0.setId("filterForm");
-      // /pages/results.jsp(176,0) name = autocomplete type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-      _jspx_th_form_005fform_005f0.setAutocomplete("off");
-      int[] _jspx_push_body_count_form_005fform_005f0 = new int[] { 0 };
-      try {
-        int _jspx_eval_form_005fform_005f0 = _jspx_th_form_005fform_005f0.doStartTag();
-        if (_jspx_eval_form_005fform_005f0 != javax.servlet.jsp.tagext.Tag.SKIP_BODY) {
-          do {
-            out.write("\r\n");
-            out.write("\r\n");
-            out.write("\t<div id=\"filterDiv\">\r\n");
-            out.write("\t\t<h2>Filter results:</h2>\r\n");
-            out.write("\t\t");
-            if (_jspx_meth_form_005fcheckbox_005f0(_jspx_th_form_005fform_005f0, _jspx_page_context, _jspx_push_body_count_form_005fform_005f0))
-              return;
-            out.write("<label>Room</label>\r\n");
-            out.write("\t\t");
-            if (_jspx_meth_form_005fcheckbox_005f1(_jspx_th_form_005fform_005f0, _jspx_page_context, _jspx_push_body_count_form_005fform_005f0))
-              return;
-            out.write("<label>Studio</label>\r\n");
-            out.write("\t\r\n");
-            out.write("\t\t");
-            if (_jspx_meth_form_005fcheckbox_005f2(_jspx_th_form_005fform_005f0, _jspx_page_context, _jspx_push_body_count_form_005fform_005f0))
-              return;
-            out.write("\r\n");
-            out.write("\t\t");
-            if (_jspx_meth_form_005fcheckbox_005f3(_jspx_th_form_005fform_005f0, _jspx_page_context, _jspx_push_body_count_form_005fform_005f0))
-              return;
-            out.write("\r\n");
-            out.write("\t\t");
-            if (_jspx_meth_form_005fcheckbox_005f4(_jspx_th_form_005fform_005f0, _jspx_page_context, _jspx_push_body_count_form_005fform_005f0))
-              return;
-            out.write("\r\n");
-            out.write("\t\t");
-            if (_jspx_meth_form_005fcheckbox_005f5(_jspx_th_form_005fform_005f0, _jspx_page_context, _jspx_push_body_count_form_005fform_005f0))
-              return;
-            out.write("\r\n");
-            out.write("\t\t");
-            //  form:errors
-            org.springframework.web.servlet.tags.form.ErrorsTag _jspx_th_form_005ferrors_005f0 = (org.springframework.web.servlet.tags.form.ErrorsTag) _005fjspx_005ftagPool_005fform_005ferrors_0026_005fpath_005fcssClass_005fnobody.get(org.springframework.web.servlet.tags.form.ErrorsTag.class);
-            _jspx_th_form_005ferrors_005f0.setPageContext(_jspx_page_context);
-            _jspx_th_form_005ferrors_005f0.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_form_005fform_005f0);
-            // /pages/results.jsp(188,2) name = path type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-            _jspx_th_form_005ferrors_005f0.setPath("noRoomNoStudio");
-            // /pages/results.jsp(188,2) name = cssClass type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-            _jspx_th_form_005ferrors_005f0.setCssClass("validationErrorText");
-            int[] _jspx_push_body_count_form_005ferrors_005f0 = new int[] { 0 };
-            try {
-              int _jspx_eval_form_005ferrors_005f0 = _jspx_th_form_005ferrors_005f0.doStartTag();
-              if (_jspx_th_form_005ferrors_005f0.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-                return;
-              }
-            } catch (java.lang.Throwable _jspx_exception) {
-              while (_jspx_push_body_count_form_005ferrors_005f0[0]-- > 0)
-                out = _jspx_page_context.popBody();
-              _jspx_th_form_005ferrors_005f0.doCatch(_jspx_exception);
-            } finally {
-              _jspx_th_form_005ferrors_005f0.doFinally();
-              _005fjspx_005ftagPool_005fform_005ferrors_0026_005fpath_005fcssClass_005fnobody.reuse(_jspx_th_form_005ferrors_005f0);
-            }
-            out.write(" <br />\r\n");
-            out.write("\t\r\n");
-            out.write("\t\t<label for=\"city\">City / zip code:</label>\r\n");
-            out.write("\t\t");
-            if (_jspx_meth_form_005finput_005f0(_jspx_th_form_005fform_005f0, _jspx_page_context, _jspx_push_body_count_form_005fform_005f0))
-              return;
-            out.write("\r\n");
-            out.write("\t\t");
-            //  form:errors
-            org.springframework.web.servlet.tags.form.ErrorsTag _jspx_th_form_005ferrors_005f1 = (org.springframework.web.servlet.tags.form.ErrorsTag) _005fjspx_005ftagPool_005fform_005ferrors_0026_005fpath_005fcssClass_005fnobody.get(org.springframework.web.servlet.tags.form.ErrorsTag.class);
-            _jspx_th_form_005ferrors_005f1.setPageContext(_jspx_page_context);
-            _jspx_th_form_005ferrors_005f1.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_form_005fform_005f0);
-            // /pages/results.jsp(193,2) name = path type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-            _jspx_th_form_005ferrors_005f1.setPath("city");
-            // /pages/results.jsp(193,2) name = cssClass type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-            _jspx_th_form_005ferrors_005f1.setCssClass("validationErrorText");
-            int[] _jspx_push_body_count_form_005ferrors_005f1 = new int[] { 0 };
-            try {
-              int _jspx_eval_form_005ferrors_005f1 = _jspx_th_form_005ferrors_005f1.doStartTag();
-              if (_jspx_th_form_005ferrors_005f1.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-                return;
-              }
-            } catch (java.lang.Throwable _jspx_exception) {
-              while (_jspx_push_body_count_form_005ferrors_005f1[0]-- > 0)
-                out = _jspx_page_context.popBody();
-              _jspx_th_form_005ferrors_005f1.doCatch(_jspx_exception);
-            } finally {
-              _jspx_th_form_005ferrors_005f1.doFinally();
-              _005fjspx_005ftagPool_005fform_005ferrors_0026_005fpath_005fcssClass_005fnobody.reuse(_jspx_th_form_005ferrors_005f1);
-            }
-            out.write("<br />\r\n");
-            out.write("\t\t\t\r\n");
-            out.write("\t\t<label for=\"radius\">Within radius of (max.):</label>\r\n");
-            out.write("\t\t");
-            if (_jspx_meth_form_005finput_005f1(_jspx_th_form_005fform_005f0, _jspx_page_context, _jspx_push_body_count_form_005fform_005f0))
-              return;
-            out.write("\r\n");
-            out.write("\t\tkm\r\n");
-            out.write("\t\t");
-            //  form:errors
-            org.springframework.web.servlet.tags.form.ErrorsTag _jspx_th_form_005ferrors_005f2 = (org.springframework.web.servlet.tags.form.ErrorsTag) _005fjspx_005ftagPool_005fform_005ferrors_0026_005fpath_005fcssClass_005fnobody.get(org.springframework.web.servlet.tags.form.ErrorsTag.class);
-            _jspx_th_form_005ferrors_005f2.setPageContext(_jspx_page_context);
-            _jspx_th_form_005ferrors_005f2.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_form_005fform_005f0);
-            // /pages/results.jsp(199,2) name = path type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-            _jspx_th_form_005ferrors_005f2.setPath("radius");
-            // /pages/results.jsp(199,2) name = cssClass type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-            _jspx_th_form_005ferrors_005f2.setCssClass("validationErrorText");
-            int[] _jspx_push_body_count_form_005ferrors_005f2 = new int[] { 0 };
-            try {
-              int _jspx_eval_form_005ferrors_005f2 = _jspx_th_form_005ferrors_005f2.doStartTag();
-              if (_jspx_th_form_005ferrors_005f2.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-                return;
-              }
-            } catch (java.lang.Throwable _jspx_exception) {
-              while (_jspx_push_body_count_form_005ferrors_005f2[0]-- > 0)
-                out = _jspx_page_context.popBody();
-              _jspx_th_form_005ferrors_005f2.doCatch(_jspx_exception);
-            } finally {
-              _jspx_th_form_005ferrors_005f2.doFinally();
-              _005fjspx_005ftagPool_005fform_005ferrors_0026_005fpath_005fcssClass_005fnobody.reuse(_jspx_th_form_005ferrors_005f2);
-            }
-            out.write("\r\n");
-            out.write("\t\t<br /> <label for=\"prize\">Price (max.):</label>\r\n");
-            out.write("\t\t");
-            if (_jspx_meth_form_005finput_005f2(_jspx_th_form_005fform_005f0, _jspx_page_context, _jspx_push_body_count_form_005fform_005f0))
-              return;
-            out.write("\r\n");
-            out.write("\t\tCHF\r\n");
-            out.write("\t\t");
-            //  form:errors
-            org.springframework.web.servlet.tags.form.ErrorsTag _jspx_th_form_005ferrors_005f3 = (org.springframework.web.servlet.tags.form.ErrorsTag) _005fjspx_005ftagPool_005fform_005ferrors_0026_005fpath_005fcssClass_005fnobody.get(org.springframework.web.servlet.tags.form.ErrorsTag.class);
-            _jspx_th_form_005ferrors_005f3.setPageContext(_jspx_page_context);
-            _jspx_th_form_005ferrors_005f3.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_form_005fform_005f0);
-            // /pages/results.jsp(204,2) name = path type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-            _jspx_th_form_005ferrors_005f3.setPath("prize");
-            // /pages/results.jsp(204,2) name = cssClass type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-            _jspx_th_form_005ferrors_005f3.setCssClass("validationErrorText");
-            int[] _jspx_push_body_count_form_005ferrors_005f3 = new int[] { 0 };
-            try {
-              int _jspx_eval_form_005ferrors_005f3 = _jspx_th_form_005ferrors_005f3.doStartTag();
-              if (_jspx_th_form_005ferrors_005f3.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-                return;
-              }
-            } catch (java.lang.Throwable _jspx_exception) {
-              while (_jspx_push_body_count_form_005ferrors_005f3[0]-- > 0)
-                out = _jspx_page_context.popBody();
-              _jspx_th_form_005ferrors_005f3.doCatch(_jspx_exception);
-            } finally {
-              _jspx_th_form_005ferrors_005f3.doFinally();
-              _005fjspx_005ftagPool_005fform_005ferrors_0026_005fpath_005fcssClass_005fnobody.reuse(_jspx_th_form_005ferrors_005f3);
-            }
-            out.write("<br />\r\n");
-            out.write("\t\t\r\n");
-            out.write("\t\t<hr class=\"slim\">\t\t\r\n");
-            out.write("\t\t\r\n");
-            out.write("\t\t<table style=\"width: 80%\">\r\n");
-            out.write("\t\t\t<tr>\r\n");
-            out.write("\t\t\t\t<td><label for=\"earliestMoveInDate\">Earliest move-in date</label></td>\r\n");
-            out.write("\t\t\t\t<td><label for=\"earliestMoveOutDate\">Earliest move-out date (optional)</label></td>\r\n");
-            out.write("\t\t\t</tr>\r\n");
-            out.write("\t\t\t<tr>\r\n");
-            out.write("\t\t\t\t<td>");
-            if (_jspx_meth_form_005finput_005f3(_jspx_th_form_005fform_005f0, _jspx_page_context, _jspx_push_body_count_form_005fform_005f0))
-              return;
-            out.write("</td>\r\n");
-            out.write("\t\t\t\t<td>");
-            if (_jspx_meth_form_005finput_005f4(_jspx_th_form_005fform_005f0, _jspx_page_context, _jspx_push_body_count_form_005fform_005f0))
-              return;
-            out.write("</td>\r\n");
-            out.write("\t\t\t</tr>\r\n");
-            out.write("\t\t\t<tr>\r\n");
-            out.write("\t\t\t\t<td><label for=\"latestMoveInDate\">Latest move-in date</label></td>\r\n");
-            out.write("\t\t\t\t<td><label for=\"latestMoveOutDate\">Latest move-out date (optional)</label></td>\r\n");
-            out.write("\t\t\t</tr>\r\n");
-            out.write("\t\t\t<tr>\r\n");
-            out.write("\t\t\t\t<td>");
-            if (_jspx_meth_form_005finput_005f5(_jspx_th_form_005fform_005f0, _jspx_page_context, _jspx_push_body_count_form_005fform_005f0))
-              return;
-            out.write("</td>\r\n");
-            out.write("\t\t\t\t<td>");
-            if (_jspx_meth_form_005finput_005f6(_jspx_th_form_005fform_005f0, _jspx_page_context, _jspx_push_body_count_form_005fform_005f0))
-              return;
-            out.write("</td>\r\n");
-            out.write("\t\t\t</tr>\r\n");
-            out.write("\t\t\t<tr>\r\n");
-            out.write("\t\t\t\t<td>");
-            if (_jspx_meth_form_005fcheckbox_005f6(_jspx_th_form_005fform_005f0, _jspx_page_context, _jspx_push_body_count_form_005fform_005f0))
-              return;
-            out.write("<label>Smoking inside\r\n");
-            out.write("\t\t\t\t\t\tallowed</label></td>\r\n");
-            out.write("\t\t\t\t<td>");
-            if (_jspx_meth_form_005fcheckbox_005f7(_jspx_th_form_005fform_005f0, _jspx_page_context, _jspx_push_body_count_form_005fform_005f0))
-              return;
-            out.write("<label>Animals\r\n");
-            out.write("\t\t\t\t\t\tinside allowed</label></td>\r\n");
-            out.write("\t\t\t</tr>\r\n");
-            out.write("\t\t\t<tr>\r\n");
-            out.write("\t\t\t\t<td>");
-            if (_jspx_meth_form_005fcheckbox_005f8(_jspx_th_form_005fform_005f0, _jspx_page_context, _jspx_push_body_count_form_005fform_005f0))
-              return;
-            out.write("<label>Garden\r\n");
-            out.write("\t\t\t\t\t\t(co-use)</label></td>\r\n");
-            out.write("\t\t\t\t<td>");
-            if (_jspx_meth_form_005fcheckbox_005f9(_jspx_th_form_005fform_005f0, _jspx_page_context, _jspx_push_body_count_form_005fform_005f0))
-              return;
-            out.write("<label>Balcony\r\n");
-            out.write("\t\t\t\t\t\tor Patio</label></td>\r\n");
-            out.write("\t\t\t</tr>\r\n");
-            out.write("\t\t\t<tr>\r\n");
-            out.write("\t\t\t\t<td>");
-            if (_jspx_meth_form_005fcheckbox_005f10(_jspx_th_form_005fform_005f0, _jspx_page_context, _jspx_push_body_count_form_005fform_005f0))
-              return;
-            out.write("<label>Cellar\r\n");
-            out.write("\t\t\t\t\t\tor Attic</label></td>\r\n");
-            out.write("\t\t\t\t<td>");
-            if (_jspx_meth_form_005fcheckbox_005f11(_jspx_th_form_005fform_005f0, _jspx_page_context, _jspx_push_body_count_form_005fform_005f0))
-              return;
-            out.write("<label>Furnished</label></td>\r\n");
-            out.write("\t\t\t</tr>\r\n");
-            out.write("\t\t\t<tr>\r\n");
-            out.write("\t\t\t\t<td>");
-            if (_jspx_meth_form_005fcheckbox_005f12(_jspx_th_form_005fform_005f0, _jspx_page_context, _jspx_push_body_count_form_005fform_005f0))
-              return;
-            out.write("<label>Cable\r\n");
-            out.write("\t\t\t\t\t\tTV</label></td>\r\n");
-            out.write("\t\t\t\t<td>");
-            if (_jspx_meth_form_005fcheckbox_005f13(_jspx_th_form_005fform_005f0, _jspx_page_context, _jspx_push_body_count_form_005fform_005f0))
-              return;
-            out.write("<label>Garage</label>\r\n");
-            out.write("\t\t\t\t</td>\r\n");
-            out.write("\t\t\t</tr>\r\n");
-            out.write("\t\t\t<tr>\r\n");
-            out.write("\t\t\t\t<td>");
-            if (_jspx_meth_form_005fcheckbox_005f14(_jspx_th_form_005fform_005f0, _jspx_page_context, _jspx_push_body_count_form_005fform_005f0))
-              return;
-            out.write("<label>WiFi</label></td>\r\n");
-            out.write("\t\t\t</tr>\r\n");
-            out.write("\t\t</table>\r\n");
-            out.write("\t\t\t\r\n");
-            out.write("\t\t\r\n");
-            out.write("\t\t<button type=\"submit\" onClick=\"validateType(this.form)\">Filter</button>\t\r\n");
-            out.write("\t\t<button type=\"reset\">Cancel</button>\r\n");
-            out.write("\t</div>\r\n");
-            int evalDoAfterBody = _jspx_th_form_005fform_005f0.doAfterBody();
-            if (evalDoAfterBody != javax.servlet.jsp.tagext.BodyTag.EVAL_BODY_AGAIN)
-              break;
-          } while (true);
-        }
-        if (_jspx_th_form_005fform_005f0.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-          return;
-        }
-      } catch (java.lang.Throwable _jspx_exception) {
-        while (_jspx_push_body_count_form_005fform_005f0[0]-- > 0)
-          out = _jspx_page_context.popBody();
-        _jspx_th_form_005fform_005f0.doCatch(_jspx_exception);
-      } finally {
-        _jspx_th_form_005fform_005f0.doFinally();
-        _005fjspx_005ftagPool_005fform_005fform_0026_005fmodelAttribute_005fmethod_005fid_005fautocomplete_005faction.reuse(_jspx_th_form_005fform_005f0);
-      }
-      out.write("\r\n");
-      out.write("\r\n");
-      if (_jspx_meth_c_005fimport_005f2(_jspx_page_context))
-        return;
-    } catch (java.lang.Throwable t) {
-      if (!(t instanceof javax.servlet.jsp.SkipPageException)){
-        out = _jspx_out;
-        if (out != null && out.getBufferSize() != 0)
-          try {
-            if (response.isCommitted()) {
-              out.flush();
-            } else {
-              out.clearBuffer();
-            }
-          } catch (java.io.IOException e) {}
-        if (_jspx_page_context != null) _jspx_page_context.handlePageException(t);
-        else throw new ServletException(t);
-      }
-    } finally {
-      _jspxFactory.releasePageContext(_jspx_page_context);
-    }
-  }
-
-  private boolean _jspx_meth_c_005fimport_005f0(javax.servlet.jsp.PageContext _jspx_page_context)
-          throws java.lang.Throwable {
-    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
-    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
-    //  c:import
-    org.apache.taglibs.standard.tag.rt.core.ImportTag _jspx_th_c_005fimport_005f0 = (org.apache.taglibs.standard.tag.rt.core.ImportTag) _005fjspx_005ftagPool_005fc_005fimport_0026_005furl_005fnobody.get(org.apache.taglibs.standard.tag.rt.core.ImportTag.class);
-    _jspx_th_c_005fimport_005f0.setPageContext(_jspx_page_context);
-    _jspx_th_c_005fimport_005f0.setParent(null);
-    // /pages/results.jsp(9,0) name = url type = null reqTime = true required = true fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_c_005fimport_005f0.setUrl("template/header.jsp");
-    int[] _jspx_push_body_count_c_005fimport_005f0 = new int[] { 0 };
-    try {
-      int _jspx_eval_c_005fimport_005f0 = _jspx_th_c_005fimport_005f0.doStartTag();
-      if (_jspx_th_c_005fimport_005f0.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-        return true;
-      }
-    } catch (java.lang.Throwable _jspx_exception) {
-      while (_jspx_push_body_count_c_005fimport_005f0[0]-- > 0)
-        out = _jspx_page_context.popBody();
-      _jspx_th_c_005fimport_005f0.doCatch(_jspx_exception);
-    } finally {
-      _jspx_th_c_005fimport_005f0.doFinally();
-      _005fjspx_005ftagPool_005fc_005fimport_0026_005furl_005fnobody.reuse(_jspx_th_c_005fimport_005f0);
-    }
-    return false;
-  }
-
-  private boolean _jspx_meth_c_005fimport_005f1(javax.servlet.jsp.PageContext _jspx_page_context)
-          throws java.lang.Throwable {
-    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
-    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
-    //  c:import
-    org.apache.taglibs.standard.tag.rt.core.ImportTag _jspx_th_c_005fimport_005f1 = (org.apache.taglibs.standard.tag.rt.core.ImportTag) _005fjspx_005ftagPool_005fc_005fimport_0026_005furl_005fnobody.get(org.apache.taglibs.standard.tag.rt.core.ImportTag.class);
-    _jspx_th_c_005fimport_005f1.setPageContext(_jspx_page_context);
-    _jspx_th_c_005fimport_005f1.setParent(null);
-    // /pages/results.jsp(98,12) name = url type = null reqTime = true required = true fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_c_005fimport_005f1.setUrl("getzipcodes.jsp");
-    int[] _jspx_push_body_count_c_005fimport_005f1 = new int[] { 0 };
-    try {
-      int _jspx_eval_c_005fimport_005f1 = _jspx_th_c_005fimport_005f1.doStartTag();
-      if (_jspx_th_c_005fimport_005f1.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-        return true;
-      }
-    } catch (java.lang.Throwable _jspx_exception) {
-      while (_jspx_push_body_count_c_005fimport_005f1[0]-- > 0)
-        out = _jspx_page_context.popBody();
-      _jspx_th_c_005fimport_005f1.doCatch(_jspx_exception);
-    } finally {
-      _jspx_th_c_005fimport_005f1.doFinally();
-      _005fjspx_005ftagPool_005fc_005fimport_0026_005furl_005fnobody.reuse(_jspx_th_c_005fimport_005f1);
-    }
-    return false;
-  }
-
-  private boolean _jspx_meth_c_005fchoose_005f0(javax.servlet.jsp.PageContext _jspx_page_context)
-          throws java.lang.Throwable {
-    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
-    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
-    //  c:choose
-    org.apache.taglibs.standard.tag.common.core.ChooseTag _jspx_th_c_005fchoose_005f0 = (org.apache.taglibs.standard.tag.common.core.ChooseTag) _005fjspx_005ftagPool_005fc_005fchoose.get(org.apache.taglibs.standard.tag.common.core.ChooseTag.class);
-    _jspx_th_c_005fchoose_005f0.setPageContext(_jspx_page_context);
-    _jspx_th_c_005fchoose_005f0.setParent(null);
-    int _jspx_eval_c_005fchoose_005f0 = _jspx_th_c_005fchoose_005f0.doStartTag();
-    if (_jspx_eval_c_005fchoose_005f0 != javax.servlet.jsp.tagext.Tag.SKIP_BODY) {
-      do {
-        out.write('\r');
-        out.write('\n');
-        out.write('	');
-        if (_jspx_meth_c_005fwhen_005f0(_jspx_th_c_005fchoose_005f0, _jspx_page_context))
-          return true;
-        out.write('\r');
-        out.write('\n');
-        out.write('	');
-        if (_jspx_meth_c_005fotherwise_005f0(_jspx_th_c_005fchoose_005f0, _jspx_page_context))
-          return true;
-        out.write('\r');
-        out.write('\n');
-        int evalDoAfterBody = _jspx_th_c_005fchoose_005f0.doAfterBody();
-        if (evalDoAfterBody != javax.servlet.jsp.tagext.BodyTag.EVAL_BODY_AGAIN)
-          break;
-      } while (true);
-    }
-    if (_jspx_th_c_005fchoose_005f0.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-      _005fjspx_005ftagPool_005fc_005fchoose.reuse(_jspx_th_c_005fchoose_005f0);
-      return true;
-    }
-    _005fjspx_005ftagPool_005fc_005fchoose.reuse(_jspx_th_c_005fchoose_005f0);
-    return false;
-  }
-
-  private boolean _jspx_meth_c_005fwhen_005f0(javax.servlet.jsp.tagext.JspTag _jspx_th_c_005fchoose_005f0, javax.servlet.jsp.PageContext _jspx_page_context)
-          throws java.lang.Throwable {
-    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
-    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
-    //  c:when
-    org.apache.taglibs.standard.tag.rt.core.WhenTag _jspx_th_c_005fwhen_005f0 = (org.apache.taglibs.standard.tag.rt.core.WhenTag) _005fjspx_005ftagPool_005fc_005fwhen_0026_005ftest.get(org.apache.taglibs.standard.tag.rt.core.WhenTag.class);
-    _jspx_th_c_005fwhen_005f0.setPageContext(_jspx_page_context);
-    _jspx_th_c_005fwhen_005f0.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_c_005fchoose_005f0);
-    // /pages/results.jsp(138,1) name = test type = boolean reqTime = true required = true fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_c_005fwhen_005f0.setTest(((java.lang.Boolean) org.apache.jasper.runtime.PageContextImpl.proprietaryEvaluate("${empty results}", boolean.class, (javax.servlet.jsp.PageContext)_jspx_page_context, null)).booleanValue());
-    int _jspx_eval_c_005fwhen_005f0 = _jspx_th_c_005fwhen_005f0.doStartTag();
-    if (_jspx_eval_c_005fwhen_005f0 != javax.servlet.jsp.tagext.Tag.SKIP_BODY) {
-      do {
-        out.write("\r\n");
-        out.write("\t\t<p>No results found!\r\n");
-        out.write("\t");
-        int evalDoAfterBody = _jspx_th_c_005fwhen_005f0.doAfterBody();
-        if (evalDoAfterBody != javax.servlet.jsp.tagext.BodyTag.EVAL_BODY_AGAIN)
-          break;
-      } while (true);
-    }
-    if (_jspx_th_c_005fwhen_005f0.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-      _005fjspx_005ftagPool_005fc_005fwhen_0026_005ftest.reuse(_jspx_th_c_005fwhen_005f0);
-      return true;
-    }
-    _005fjspx_005ftagPool_005fc_005fwhen_0026_005ftest.reuse(_jspx_th_c_005fwhen_005f0);
-    return false;
-  }
-
-  private boolean _jspx_meth_c_005fotherwise_005f0(javax.servlet.jsp.tagext.JspTag _jspx_th_c_005fchoose_005f0, javax.servlet.jsp.PageContext _jspx_page_context)
-          throws java.lang.Throwable {
-    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
-    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
-    //  c:otherwise
-    org.apache.taglibs.standard.tag.common.core.OtherwiseTag _jspx_th_c_005fotherwise_005f0 = (org.apache.taglibs.standard.tag.common.core.OtherwiseTag) _005fjspx_005ftagPool_005fc_005fotherwise.get(org.apache.taglibs.standard.tag.common.core.OtherwiseTag.class);
-    _jspx_th_c_005fotherwise_005f0.setPageContext(_jspx_page_context);
-    _jspx_th_c_005fotherwise_005f0.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_c_005fchoose_005f0);
-    int _jspx_eval_c_005fotherwise_005f0 = _jspx_th_c_005fotherwise_005f0.doStartTag();
-    if (_jspx_eval_c_005fotherwise_005f0 != javax.servlet.jsp.tagext.Tag.SKIP_BODY) {
-      do {
-        out.write("\r\n");
-        out.write("\t\t<div id=\"resultsDiv\" class=\"resultsDiv\">\t\t\t\r\n");
-        out.write("\t\t\t");
-        if (_jspx_meth_c_005fforEach_005f0(_jspx_th_c_005fotherwise_005f0, _jspx_page_context))
-          return true;
-        out.write("\r\n");
-        out.write("\t\t</div>\r\n");
-        out.write("\t");
-        int evalDoAfterBody = _jspx_th_c_005fotherwise_005f0.doAfterBody();
-        if (evalDoAfterBody != javax.servlet.jsp.tagext.BodyTag.EVAL_BODY_AGAIN)
-          break;
-      } while (true);
-    }
-    if (_jspx_th_c_005fotherwise_005f0.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-      _005fjspx_005ftagPool_005fc_005fotherwise.reuse(_jspx_th_c_005fotherwise_005f0);
-      return true;
-    }
-    _005fjspx_005ftagPool_005fc_005fotherwise.reuse(_jspx_th_c_005fotherwise_005f0);
-    return false;
-  }
-
-  private boolean _jspx_meth_c_005fforEach_005f0(javax.servlet.jsp.tagext.JspTag _jspx_th_c_005fotherwise_005f0, javax.servlet.jsp.PageContext _jspx_page_context)
-          throws java.lang.Throwable {
-    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
-    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
-    //  c:forEach
-    org.apache.taglibs.standard.tag.rt.core.ForEachTag _jspx_th_c_005fforEach_005f0 = (org.apache.taglibs.standard.tag.rt.core.ForEachTag) _005fjspx_005ftagPool_005fc_005fforEach_0026_005fvar_005fitems.get(org.apache.taglibs.standard.tag.rt.core.ForEachTag.class);
-    _jspx_th_c_005fforEach_005f0.setPageContext(_jspx_page_context);
-    _jspx_th_c_005fforEach_005f0.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_c_005fotherwise_005f0);
-    // /pages/results.jsp(143,3) name = var type = java.lang.String reqTime = false required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_c_005fforEach_005f0.setVar("ad");
-    // /pages/results.jsp(143,3) name = items type = javax.el.ValueExpression reqTime = true required = false fragment = false deferredValue = true expectedTypeName = java.lang.Object deferredMethod = false methodSignature = null
-    _jspx_th_c_005fforEach_005f0.setItems(new org.apache.jasper.el.JspValueExpression("/pages/results.jsp(143,3) '${results}'",_el_expressionfactory.createValueExpression(_jspx_page_context.getELContext(),"${results}",java.lang.Object.class)).getValue(_jspx_page_context.getELContext()));
-    int[] _jspx_push_body_count_c_005fforEach_005f0 = new int[] { 0 };
-    try {
-      int _jspx_eval_c_005fforEach_005f0 = _jspx_th_c_005fforEach_005f0.doStartTag();
-      if (_jspx_eval_c_005fforEach_005f0 != javax.servlet.jsp.tagext.Tag.SKIP_BODY) {
-        do {
-          out.write("\r\n");
-          out.write("\t\t\t\t<div class=\"resultAd\" data-price=\"");
-          out.write((java.lang.String) org.apache.jasper.runtime.PageContextImpl.proprietaryEvaluate("${ad.prizePerMonth}", java.lang.String.class, (javax.servlet.jsp.PageContext)_jspx_page_context, null));
-          out.write("\" \r\n");
-          out.write("\t\t\t\t\t\t\t\tdata-moveIn=\"");
-          out.write((java.lang.String) org.apache.jasper.runtime.PageContextImpl.proprietaryEvaluate("${ad.moveInDate}", java.lang.String.class, (javax.servlet.jsp.PageContext)_jspx_page_context, null));
-          out.write("\" data-age=\"");
-          out.write((java.lang.String) org.apache.jasper.runtime.PageContextImpl.proprietaryEvaluate("${ad.moveInDate}", java.lang.String.class, (javax.servlet.jsp.PageContext)_jspx_page_context, null));
-          out.write("\">\r\n");
-          out.write("\t\t\t\t\t<div class=\"resultLeft\">\r\n");
-          out.write("\t\t\t\t\t\t<a href=\"");
-          if (_jspx_meth_c_005furl_005f0(_jspx_th_c_005fforEach_005f0, _jspx_page_context, _jspx_push_body_count_c_005fforEach_005f0))
-            return true;
-          out.write("\"><img\r\n");
-          out.write("\t\t\t\t\t\t\tsrc=\"");
-          out.write((java.lang.String) org.apache.jasper.runtime.PageContextImpl.proprietaryEvaluate("${ad.pictures[0].filePath}", java.lang.String.class, (javax.servlet.jsp.PageContext)_jspx_page_context, null));
-          out.write("\" /></a>\r\n");
-          out.write("\t\t\t\t\t\t<h2>\r\n");
-          out.write("\t\t\t\t\t\t\t<a class=\"link\" href=\"");
-          if (_jspx_meth_c_005furl_005f1(_jspx_th_c_005fforEach_005f0, _jspx_page_context, _jspx_push_body_count_c_005fforEach_005f0))
-            return true;
-          out.write('"');
-          out.write('>');
-          out.write((java.lang.String) org.apache.jasper.runtime.PageContextImpl.proprietaryEvaluate("${ad.title }", java.lang.String.class, (javax.servlet.jsp.PageContext)_jspx_page_context, null));
-          out.write("</a>\r\n");
-          out.write("\t\t\t\t\t\t</h2>\r\n");
-          out.write("\t\t\t\t\t\t<p>");
-          out.write((java.lang.String) org.apache.jasper.runtime.PageContextImpl.proprietaryEvaluate("${ad.street}", java.lang.String.class, (javax.servlet.jsp.PageContext)_jspx_page_context, null));
-          out.write(',');
-          out.write(' ');
-          out.write((java.lang.String) org.apache.jasper.runtime.PageContextImpl.proprietaryEvaluate("${ad.zipcode}", java.lang.String.class, (javax.servlet.jsp.PageContext)_jspx_page_context, null));
-          out.write(' ');
-          out.write((java.lang.String) org.apache.jasper.runtime.PageContextImpl.proprietaryEvaluate("${ad.city}", java.lang.String.class, (javax.servlet.jsp.PageContext)_jspx_page_context, null));
-          out.write("</p>\r\n");
-          out.write("\t\t\t\t\t\t<br />\r\n");
-          out.write("\t\t\t\t\t\t<p>\r\n");
-          out.write("\t\t\t\t\t\t\t<i>");
-          if (_jspx_meth_c_005fchoose_005f1(_jspx_th_c_005fforEach_005f0, _jspx_page_context, _jspx_push_body_count_c_005fforEach_005f0))
-            return true;
-          out.write("</i>\r\n");
-          out.write("\t\t\t\t\t\t</p>\r\n");
-          out.write("\t\t\t\t\t</div>\r\n");
-          out.write("\t\t\t\t\t<div class=\"resultRight\">\r\n");
-          out.write("\t\t\t\t\t\t<h2>CHF ");
-          out.write((java.lang.String) org.apache.jasper.runtime.PageContextImpl.proprietaryEvaluate("${ad.prizePerMonth }", java.lang.String.class, (javax.servlet.jsp.PageContext)_jspx_page_context, null));
-          out.write("</h2>\r\n");
-          out.write("\t\t\t\t\t\t<br /> <br />\r\n");
-          out.write("\r\n");
-          out.write("\t\t\t\t\t\t");
-          if (_jspx_meth_fmt_005fformatDate_005f0(_jspx_th_c_005fforEach_005f0, _jspx_page_context, _jspx_push_body_count_c_005fforEach_005f0))
-            return true;
-          out.write("\r\n");
-          out.write("\r\n");
-          out.write("\t\t\t\t\t\t<p>Move-in date: ");
-          out.write((java.lang.String) org.apache.jasper.runtime.PageContextImpl.proprietaryEvaluate("${formattedMoveInDate }", java.lang.String.class, (javax.servlet.jsp.PageContext)_jspx_page_context, null));
-          out.write("</p>\r\n");
-          out.write("\t\t\t\t\t</div>\r\n");
-          out.write("\t\t\t\t</div>\r\n");
-          out.write("\t\t\t");
-          int evalDoAfterBody = _jspx_th_c_005fforEach_005f0.doAfterBody();
-          if (evalDoAfterBody != javax.servlet.jsp.tagext.BodyTag.EVAL_BODY_AGAIN)
-            break;
-        } while (true);
-      }
-      if (_jspx_th_c_005fforEach_005f0.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-        return true;
-      }
-    } catch (java.lang.Throwable _jspx_exception) {
-      while (_jspx_push_body_count_c_005fforEach_005f0[0]-- > 0)
-        out = _jspx_page_context.popBody();
-      _jspx_th_c_005fforEach_005f0.doCatch(_jspx_exception);
-    } finally {
-      _jspx_th_c_005fforEach_005f0.doFinally();
-      _005fjspx_005ftagPool_005fc_005fforEach_0026_005fvar_005fitems.reuse(_jspx_th_c_005fforEach_005f0);
-    }
-    return false;
-  }
-
-  private boolean _jspx_meth_c_005furl_005f0(javax.servlet.jsp.tagext.JspTag _jspx_th_c_005fforEach_005f0, javax.servlet.jsp.PageContext _jspx_page_context, int[] _jspx_push_body_count_c_005fforEach_005f0)
-          throws java.lang.Throwable {
-    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
-    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
-    //  c:url
-    org.apache.taglibs.standard.tag.rt.core.UrlTag _jspx_th_c_005furl_005f0 = (org.apache.taglibs.standard.tag.rt.core.UrlTag) _005fjspx_005ftagPool_005fc_005furl_0026_005fvalue_005fnobody.get(org.apache.taglibs.standard.tag.rt.core.UrlTag.class);
-    _jspx_th_c_005furl_005f0.setPageContext(_jspx_page_context);
-    _jspx_th_c_005furl_005f0.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_c_005fforEach_005f0);
-    // /pages/results.jsp(147,15) name = value type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_c_005furl_005f0.setValue((java.lang.String) org.apache.jasper.runtime.PageContextImpl.proprietaryEvaluate("/ad?id=${ad.id}", java.lang.String.class, (javax.servlet.jsp.PageContext)_jspx_page_context, null));
-    int _jspx_eval_c_005furl_005f0 = _jspx_th_c_005furl_005f0.doStartTag();
-    if (_jspx_th_c_005furl_005f0.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-      _005fjspx_005ftagPool_005fc_005furl_0026_005fvalue_005fnobody.reuse(_jspx_th_c_005furl_005f0);
-      return true;
-    }
-    _005fjspx_005ftagPool_005fc_005furl_0026_005fvalue_005fnobody.reuse(_jspx_th_c_005furl_005f0);
-    return false;
-  }
-
-  private boolean _jspx_meth_c_005furl_005f1(javax.servlet.jsp.tagext.JspTag _jspx_th_c_005fforEach_005f0, javax.servlet.jsp.PageContext _jspx_page_context, int[] _jspx_push_body_count_c_005fforEach_005f0)
-          throws java.lang.Throwable {
-    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
-    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
-    //  c:url
-    org.apache.taglibs.standard.tag.rt.core.UrlTag _jspx_th_c_005furl_005f1 = (org.apache.taglibs.standard.tag.rt.core.UrlTag) _005fjspx_005ftagPool_005fc_005furl_0026_005fvalue_005fnobody.get(org.apache.taglibs.standard.tag.rt.core.UrlTag.class);
-    _jspx_th_c_005furl_005f1.setPageContext(_jspx_page_context);
-    _jspx_th_c_005furl_005f1.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_c_005fforEach_005f0);
-    // /pages/results.jsp(150,29) name = value type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_c_005furl_005f1.setValue((java.lang.String) org.apache.jasper.runtime.PageContextImpl.proprietaryEvaluate("/ad?id=${ad.id}", java.lang.String.class, (javax.servlet.jsp.PageContext)_jspx_page_context, null));
-    int _jspx_eval_c_005furl_005f1 = _jspx_th_c_005furl_005f1.doStartTag();
-    if (_jspx_th_c_005furl_005f1.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-      _005fjspx_005ftagPool_005fc_005furl_0026_005fvalue_005fnobody.reuse(_jspx_th_c_005furl_005f1);
-      return true;
-    }
-    _005fjspx_005ftagPool_005fc_005furl_0026_005fvalue_005fnobody.reuse(_jspx_th_c_005furl_005f1);
-    return false;
-  }
-
-  private boolean _jspx_meth_c_005fchoose_005f1(javax.servlet.jsp.tagext.JspTag _jspx_th_c_005fforEach_005f0, javax.servlet.jsp.PageContext _jspx_page_context, int[] _jspx_push_body_count_c_005fforEach_005f0)
-          throws java.lang.Throwable {
-    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
-    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
-    //  c:choose
-    org.apache.taglibs.standard.tag.common.core.ChooseTag _jspx_th_c_005fchoose_005f1 = (org.apache.taglibs.standard.tag.common.core.ChooseTag) _005fjspx_005ftagPool_005fc_005fchoose.get(org.apache.taglibs.standard.tag.common.core.ChooseTag.class);
-    _jspx_th_c_005fchoose_005f1.setPageContext(_jspx_page_context);
-    _jspx_th_c_005fchoose_005f1.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_c_005fforEach_005f0);
-    int _jspx_eval_c_005fchoose_005f1 = _jspx_th_c_005fchoose_005f1.doStartTag();
-    if (_jspx_eval_c_005fchoose_005f1 != javax.servlet.jsp.tagext.Tag.SKIP_BODY) {
-      do {
-        out.write("\r\n");
-        out.write("\t\t\t\t\t\t\t\t\t");
-        if (_jspx_meth_c_005fwhen_005f1(_jspx_th_c_005fchoose_005f1, _jspx_page_context, _jspx_push_body_count_c_005fforEach_005f0))
-          return true;
-        out.write("\r\n");
-        out.write("\t\t\t\t\t\t\t\t\t");
-        if (_jspx_meth_c_005fotherwise_005f1(_jspx_th_c_005fchoose_005f1, _jspx_page_context, _jspx_push_body_count_c_005fforEach_005f0))
-          return true;
-        out.write("\r\n");
-        out.write("\t\t\t\t\t\t\t\t");
-        int evalDoAfterBody = _jspx_th_c_005fchoose_005f1.doAfterBody();
-        if (evalDoAfterBody != javax.servlet.jsp.tagext.BodyTag.EVAL_BODY_AGAIN)
-          break;
-      } while (true);
-    }
-    if (_jspx_th_c_005fchoose_005f1.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-      _005fjspx_005ftagPool_005fc_005fchoose.reuse(_jspx_th_c_005fchoose_005f1);
-      return true;
-    }
-    _005fjspx_005ftagPool_005fc_005fchoose.reuse(_jspx_th_c_005fchoose_005f1);
-    return false;
-  }
-
-  private boolean _jspx_meth_c_005fwhen_005f1(javax.servlet.jsp.tagext.JspTag _jspx_th_c_005fchoose_005f1, javax.servlet.jsp.PageContext _jspx_page_context, int[] _jspx_push_body_count_c_005fforEach_005f0)
-          throws java.lang.Throwable {
-    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
-    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
-    //  c:when
-    org.apache.taglibs.standard.tag.rt.core.WhenTag _jspx_th_c_005fwhen_005f1 = (org.apache.taglibs.standard.tag.rt.core.WhenTag) _005fjspx_005ftagPool_005fc_005fwhen_0026_005ftest.get(org.apache.taglibs.standard.tag.rt.core.WhenTag.class);
-    _jspx_th_c_005fwhen_005f1.setPageContext(_jspx_page_context);
-    _jspx_th_c_005fwhen_005f1.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_c_005fchoose_005f1);
-    // /pages/results.jsp(156,9) name = test type = boolean reqTime = true required = true fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_c_005fwhen_005f1.setTest(((java.lang.Boolean) org.apache.jasper.runtime.PageContextImpl.proprietaryEvaluate("${ad.studio}", boolean.class, (javax.servlet.jsp.PageContext)_jspx_page_context, null)).booleanValue());
-    int _jspx_eval_c_005fwhen_005f1 = _jspx_th_c_005fwhen_005f1.doStartTag();
-    if (_jspx_eval_c_005fwhen_005f1 != javax.servlet.jsp.tagext.Tag.SKIP_BODY) {
-      do {
-        out.write("Studio");
-        int evalDoAfterBody = _jspx_th_c_005fwhen_005f1.doAfterBody();
-        if (evalDoAfterBody != javax.servlet.jsp.tagext.BodyTag.EVAL_BODY_AGAIN)
-          break;
-      } while (true);
-    }
-    if (_jspx_th_c_005fwhen_005f1.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-      _005fjspx_005ftagPool_005fc_005fwhen_0026_005ftest.reuse(_jspx_th_c_005fwhen_005f1);
-      return true;
-    }
-    _005fjspx_005ftagPool_005fc_005fwhen_0026_005ftest.reuse(_jspx_th_c_005fwhen_005f1);
-    return false;
-  }
-
-  private boolean _jspx_meth_c_005fotherwise_005f1(javax.servlet.jsp.tagext.JspTag _jspx_th_c_005fchoose_005f1, javax.servlet.jsp.PageContext _jspx_page_context, int[] _jspx_push_body_count_c_005fforEach_005f0)
-          throws java.lang.Throwable {
-    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
-    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
-    //  c:otherwise
-    org.apache.taglibs.standard.tag.common.core.OtherwiseTag _jspx_th_c_005fotherwise_005f1 = (org.apache.taglibs.standard.tag.common.core.OtherwiseTag) _005fjspx_005ftagPool_005fc_005fotherwise.get(org.apache.taglibs.standard.tag.common.core.OtherwiseTag.class);
-    _jspx_th_c_005fotherwise_005f1.setPageContext(_jspx_page_context);
-    _jspx_th_c_005fotherwise_005f1.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_c_005fchoose_005f1);
-    int _jspx_eval_c_005fotherwise_005f1 = _jspx_th_c_005fotherwise_005f1.doStartTag();
-    if (_jspx_eval_c_005fotherwise_005f1 != javax.servlet.jsp.tagext.Tag.SKIP_BODY) {
-      do {
-        out.write("Room");
-        int evalDoAfterBody = _jspx_th_c_005fotherwise_005f1.doAfterBody();
-        if (evalDoAfterBody != javax.servlet.jsp.tagext.BodyTag.EVAL_BODY_AGAIN)
-          break;
-      } while (true);
-    }
-    if (_jspx_th_c_005fotherwise_005f1.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-      _005fjspx_005ftagPool_005fc_005fotherwise.reuse(_jspx_th_c_005fotherwise_005f1);
-      return true;
-    }
-    _005fjspx_005ftagPool_005fc_005fotherwise.reuse(_jspx_th_c_005fotherwise_005f1);
-    return false;
-  }
-
-  private boolean _jspx_meth_fmt_005fformatDate_005f0(javax.servlet.jsp.tagext.JspTag _jspx_th_c_005fforEach_005f0, javax.servlet.jsp.PageContext _jspx_page_context, int[] _jspx_push_body_count_c_005fforEach_005f0)
-          throws java.lang.Throwable {
-    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
-    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
-    //  fmt:formatDate
-    org.apache.taglibs.standard.tag.rt.fmt.FormatDateTag _jspx_th_fmt_005fformatDate_005f0 = (org.apache.taglibs.standard.tag.rt.fmt.FormatDateTag) _005fjspx_005ftagPool_005ffmt_005fformatDate_0026_005fvar_005fvalue_005ftype_005fpattern_005fnobody.get(org.apache.taglibs.standard.tag.rt.fmt.FormatDateTag.class);
-    _jspx_th_fmt_005fformatDate_005f0.setPageContext(_jspx_page_context);
-    _jspx_th_fmt_005fformatDate_005f0.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_c_005fforEach_005f0);
-    // /pages/results.jsp(165,6) name = value type = null reqTime = true required = true fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_fmt_005fformatDate_005f0.setValue((java.util.Date) org.apache.jasper.runtime.PageContextImpl.proprietaryEvaluate("${ad.moveInDate}", java.util.Date.class, (javax.servlet.jsp.PageContext)_jspx_page_context, null));
-    // /pages/results.jsp(165,6) name = var type = java.lang.String reqTime = false required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_fmt_005fformatDate_005f0.setVar("formattedMoveInDate");
-    // /pages/results.jsp(165,6) name = type type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_fmt_005fformatDate_005f0.setType("date");
-    // /pages/results.jsp(165,6) name = pattern type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_fmt_005fformatDate_005f0.setPattern("dd.MM.yyyy");
-    int _jspx_eval_fmt_005fformatDate_005f0 = _jspx_th_fmt_005fformatDate_005f0.doStartTag();
-    if (_jspx_th_fmt_005fformatDate_005f0.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-      _005fjspx_005ftagPool_005ffmt_005fformatDate_0026_005fvar_005fvalue_005ftype_005fpattern_005fnobody.reuse(_jspx_th_fmt_005fformatDate_005f0);
-      return true;
-    }
-    _005fjspx_005ftagPool_005ffmt_005fformatDate_0026_005fvar_005fvalue_005ftype_005fpattern_005fnobody.reuse(_jspx_th_fmt_005fformatDate_005f0);
-    return false;
-  }
-
-  private boolean _jspx_meth_form_005fcheckbox_005f0(javax.servlet.jsp.tagext.JspTag _jspx_th_form_005fform_005f0, javax.servlet.jsp.PageContext _jspx_page_context, int[] _jspx_push_body_count_form_005fform_005f0)
-          throws java.lang.Throwable {
-    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
-    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
-    //  form:checkbox
-    org.springframework.web.servlet.tags.form.CheckboxTag _jspx_th_form_005fcheckbox_005f0 = (org.springframework.web.servlet.tags.form.CheckboxTag) _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fpath_005fname_005fid_005fnobody.get(org.springframework.web.servlet.tags.form.CheckboxTag.class);
-    _jspx_th_form_005fcheckbox_005f0.setPageContext(_jspx_page_context);
-    _jspx_th_form_005fcheckbox_005f0.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_form_005fform_005f0);
-    // /pages/results.jsp(181,2) null
-    _jspx_th_form_005fcheckbox_005f0.setDynamicAttribute(null, "name", "room");
-    // /pages/results.jsp(181,2) name = id type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005fcheckbox_005f0.setId("room");
-    // /pages/results.jsp(181,2) name = path type = null reqTime = true required = true fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005fcheckbox_005f0.setPath("roomHelper");
-    int[] _jspx_push_body_count_form_005fcheckbox_005f0 = new int[] { 0 };
-    try {
-      int _jspx_eval_form_005fcheckbox_005f0 = _jspx_th_form_005fcheckbox_005f0.doStartTag();
-      if (_jspx_th_form_005fcheckbox_005f0.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-        return true;
-      }
-    } catch (java.lang.Throwable _jspx_exception) {
-      while (_jspx_push_body_count_form_005fcheckbox_005f0[0]-- > 0)
-        out = _jspx_page_context.popBody();
-      _jspx_th_form_005fcheckbox_005f0.doCatch(_jspx_exception);
-    } finally {
-      _jspx_th_form_005fcheckbox_005f0.doFinally();
-      _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fpath_005fname_005fid_005fnobody.reuse(_jspx_th_form_005fcheckbox_005f0);
-    }
-    return false;
-  }
-
-  private boolean _jspx_meth_form_005fcheckbox_005f1(javax.servlet.jsp.tagext.JspTag _jspx_th_form_005fform_005f0, javax.servlet.jsp.PageContext _jspx_page_context, int[] _jspx_push_body_count_form_005fform_005f0)
-          throws java.lang.Throwable {
-    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
-    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
-    //  form:checkbox
-    org.springframework.web.servlet.tags.form.CheckboxTag _jspx_th_form_005fcheckbox_005f1 = (org.springframework.web.servlet.tags.form.CheckboxTag) _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fpath_005fname_005fid_005fnobody.get(org.springframework.web.servlet.tags.form.CheckboxTag.class);
-    _jspx_th_form_005fcheckbox_005f1.setPageContext(_jspx_page_context);
-    _jspx_th_form_005fcheckbox_005f1.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_form_005fform_005f0);
-    // /pages/results.jsp(182,2) null
-    _jspx_th_form_005fcheckbox_005f1.setDynamicAttribute(null, "name", "studio");
-    // /pages/results.jsp(182,2) name = id type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005fcheckbox_005f1.setId("studio");
-    // /pages/results.jsp(182,2) name = path type = null reqTime = true required = true fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005fcheckbox_005f1.setPath("studioHelper");
-    int[] _jspx_push_body_count_form_005fcheckbox_005f1 = new int[] { 0 };
-    try {
-      int _jspx_eval_form_005fcheckbox_005f1 = _jspx_th_form_005fcheckbox_005f1.doStartTag();
-      if (_jspx_th_form_005fcheckbox_005f1.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-        return true;
-      }
-    } catch (java.lang.Throwable _jspx_exception) {
-      while (_jspx_push_body_count_form_005fcheckbox_005f1[0]-- > 0)
-        out = _jspx_page_context.popBody();
-      _jspx_th_form_005fcheckbox_005f1.doCatch(_jspx_exception);
-    } finally {
-      _jspx_th_form_005fcheckbox_005f1.doFinally();
-      _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fpath_005fname_005fid_005fnobody.reuse(_jspx_th_form_005fcheckbox_005f1);
-    }
-    return false;
-  }
-
-  private boolean _jspx_meth_form_005fcheckbox_005f2(javax.servlet.jsp.tagext.JspTag _jspx_th_form_005fform_005f0, javax.servlet.jsp.PageContext _jspx_page_context, int[] _jspx_push_body_count_form_005fform_005f0)
-          throws java.lang.Throwable {
-    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
-    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
-    //  form:checkbox
-    org.springframework.web.servlet.tags.form.CheckboxTag _jspx_th_form_005fcheckbox_005f2 = (org.springframework.web.servlet.tags.form.CheckboxTag) _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fstyle_005fpath_005fname_005fid_005fnobody.get(org.springframework.web.servlet.tags.form.CheckboxTag.class);
-    _jspx_th_form_005fcheckbox_005f2.setPageContext(_jspx_page_context);
-    _jspx_th_form_005fcheckbox_005f2.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_form_005fform_005f0);
-    // /pages/results.jsp(184,2) null
-    _jspx_th_form_005fcheckbox_005f2.setDynamicAttribute(null, "style", "display:none");
-    // /pages/results.jsp(184,2) null
-    _jspx_th_form_005fcheckbox_005f2.setDynamicAttribute(null, "name", "neither");
-    // /pages/results.jsp(184,2) name = id type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005fcheckbox_005f2.setId("neither");
-    // /pages/results.jsp(184,2) name = path type = null reqTime = true required = true fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005fcheckbox_005f2.setPath("noRoomNoStudio");
-    int[] _jspx_push_body_count_form_005fcheckbox_005f2 = new int[] { 0 };
-    try {
-      int _jspx_eval_form_005fcheckbox_005f2 = _jspx_th_form_005fcheckbox_005f2.doStartTag();
-      if (_jspx_th_form_005fcheckbox_005f2.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-        return true;
-      }
-    } catch (java.lang.Throwable _jspx_exception) {
-      while (_jspx_push_body_count_form_005fcheckbox_005f2[0]-- > 0)
-        out = _jspx_page_context.popBody();
-      _jspx_th_form_005fcheckbox_005f2.doCatch(_jspx_exception);
-    } finally {
-      _jspx_th_form_005fcheckbox_005f2.doFinally();
-      _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fstyle_005fpath_005fname_005fid_005fnobody.reuse(_jspx_th_form_005fcheckbox_005f2);
-    }
-    return false;
-  }
-
-  private boolean _jspx_meth_form_005fcheckbox_005f3(javax.servlet.jsp.tagext.JspTag _jspx_th_form_005fform_005f0, javax.servlet.jsp.PageContext _jspx_page_context, int[] _jspx_push_body_count_form_005fform_005f0)
-          throws java.lang.Throwable {
-    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
-    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
-    //  form:checkbox
-    org.springframework.web.servlet.tags.form.CheckboxTag _jspx_th_form_005fcheckbox_005f3 = (org.springframework.web.servlet.tags.form.CheckboxTag) _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fstyle_005fpath_005fname_005fid_005fnobody.get(org.springframework.web.servlet.tags.form.CheckboxTag.class);
-    _jspx_th_form_005fcheckbox_005f3.setPageContext(_jspx_page_context);
-    _jspx_th_form_005fcheckbox_005f3.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_form_005fform_005f0);
-    // /pages/results.jsp(185,2) null
-    _jspx_th_form_005fcheckbox_005f3.setDynamicAttribute(null, "style", "display:none");
-    // /pages/results.jsp(185,2) null
-    _jspx_th_form_005fcheckbox_005f3.setDynamicAttribute(null, "name", "both");
-    // /pages/results.jsp(185,2) name = id type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005fcheckbox_005f3.setId("both");
-    // /pages/results.jsp(185,2) name = path type = null reqTime = true required = true fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005fcheckbox_005f3.setPath("bothRoomAndStudio");
-    int[] _jspx_push_body_count_form_005fcheckbox_005f3 = new int[] { 0 };
-    try {
-      int _jspx_eval_form_005fcheckbox_005f3 = _jspx_th_form_005fcheckbox_005f3.doStartTag();
-      if (_jspx_th_form_005fcheckbox_005f3.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-        return true;
-      }
-    } catch (java.lang.Throwable _jspx_exception) {
-      while (_jspx_push_body_count_form_005fcheckbox_005f3[0]-- > 0)
-        out = _jspx_page_context.popBody();
-      _jspx_th_form_005fcheckbox_005f3.doCatch(_jspx_exception);
-    } finally {
-      _jspx_th_form_005fcheckbox_005f3.doFinally();
-      _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fstyle_005fpath_005fname_005fid_005fnobody.reuse(_jspx_th_form_005fcheckbox_005f3);
-    }
-    return false;
-  }
-
-  private boolean _jspx_meth_form_005fcheckbox_005f4(javax.servlet.jsp.tagext.JspTag _jspx_th_form_005fform_005f0, javax.servlet.jsp.PageContext _jspx_page_context, int[] _jspx_push_body_count_form_005fform_005f0)
-          throws java.lang.Throwable {
-    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
-    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
-    //  form:checkbox
-    org.springframework.web.servlet.tags.form.CheckboxTag _jspx_th_form_005fcheckbox_005f4 = (org.springframework.web.servlet.tags.form.CheckboxTag) _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fstyle_005fpath_005fname_005fid_005fnobody.get(org.springframework.web.servlet.tags.form.CheckboxTag.class);
-    _jspx_th_form_005fcheckbox_005f4.setPageContext(_jspx_page_context);
-    _jspx_th_form_005fcheckbox_005f4.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_form_005fform_005f0);
-    // /pages/results.jsp(186,2) null
-    _jspx_th_form_005fcheckbox_005f4.setDynamicAttribute(null, "style", "display:none");
-    // /pages/results.jsp(186,2) null
-    _jspx_th_form_005fcheckbox_005f4.setDynamicAttribute(null, "name", "type");
-    // /pages/results.jsp(186,2) name = id type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005fcheckbox_005f4.setId("type");
-    // /pages/results.jsp(186,2) name = path type = null reqTime = true required = true fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005fcheckbox_005f4.setPath("studio");
-    int[] _jspx_push_body_count_form_005fcheckbox_005f4 = new int[] { 0 };
-    try {
-      int _jspx_eval_form_005fcheckbox_005f4 = _jspx_th_form_005fcheckbox_005f4.doStartTag();
-      if (_jspx_th_form_005fcheckbox_005f4.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-        return true;
-      }
-    } catch (java.lang.Throwable _jspx_exception) {
-      while (_jspx_push_body_count_form_005fcheckbox_005f4[0]-- > 0)
-        out = _jspx_page_context.popBody();
-      _jspx_th_form_005fcheckbox_005f4.doCatch(_jspx_exception);
-    } finally {
-      _jspx_th_form_005fcheckbox_005f4.doFinally();
-      _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fstyle_005fpath_005fname_005fid_005fnobody.reuse(_jspx_th_form_005fcheckbox_005f4);
-    }
-    return false;
-  }
-
-  private boolean _jspx_meth_form_005fcheckbox_005f5(javax.servlet.jsp.tagext.JspTag _jspx_th_form_005fform_005f0, javax.servlet.jsp.PageContext _jspx_page_context, int[] _jspx_push_body_count_form_005fform_005f0)
-          throws java.lang.Throwable {
-    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
-    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
-    //  form:checkbox
-    org.springframework.web.servlet.tags.form.CheckboxTag _jspx_th_form_005fcheckbox_005f5 = (org.springframework.web.servlet.tags.form.CheckboxTag) _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fstyle_005fpath_005fname_005fid_005fnobody.get(org.springframework.web.servlet.tags.form.CheckboxTag.class);
-    _jspx_th_form_005fcheckbox_005f5.setPageContext(_jspx_page_context);
-    _jspx_th_form_005fcheckbox_005f5.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_form_005fform_005f0);
-    // /pages/results.jsp(187,2) null
-    _jspx_th_form_005fcheckbox_005f5.setDynamicAttribute(null, "style", "display:none");
-    // /pages/results.jsp(187,2) null
-    _jspx_th_form_005fcheckbox_005f5.setDynamicAttribute(null, "name", "filtered");
-    // /pages/results.jsp(187,2) name = id type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005fcheckbox_005f5.setId("filtered");
-    // /pages/results.jsp(187,2) name = path type = null reqTime = true required = true fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005fcheckbox_005f5.setPath("filtered");
-    int[] _jspx_push_body_count_form_005fcheckbox_005f5 = new int[] { 0 };
-    try {
-      int _jspx_eval_form_005fcheckbox_005f5 = _jspx_th_form_005fcheckbox_005f5.doStartTag();
-      if (_jspx_th_form_005fcheckbox_005f5.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-        return true;
-      }
-    } catch (java.lang.Throwable _jspx_exception) {
-      while (_jspx_push_body_count_form_005fcheckbox_005f5[0]-- > 0)
-        out = _jspx_page_context.popBody();
-      _jspx_th_form_005fcheckbox_005f5.doCatch(_jspx_exception);
-    } finally {
-      _jspx_th_form_005fcheckbox_005f5.doFinally();
-      _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fstyle_005fpath_005fname_005fid_005fnobody.reuse(_jspx_th_form_005fcheckbox_005f5);
-    }
-    return false;
-  }
-
-  private boolean _jspx_meth_form_005finput_005f0(javax.servlet.jsp.tagext.JspTag _jspx_th_form_005fform_005f0, javax.servlet.jsp.PageContext _jspx_page_context, int[] _jspx_push_body_count_form_005fform_005f0)
-          throws java.lang.Throwable {
-    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
-    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
-    //  form:input
-    org.springframework.web.servlet.tags.form.InputTag _jspx_th_form_005finput_005f0 = (org.springframework.web.servlet.tags.form.InputTag) _005fjspx_005ftagPool_005fform_005finput_0026_005ftype_005ftabindex_005fplaceholder_005fpath_005fname_005fid_005fnobody.get(org.springframework.web.servlet.tags.form.InputTag.class);
-    _jspx_th_form_005finput_005f0.setPageContext(_jspx_page_context);
-    _jspx_th_form_005finput_005f0.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_form_005fform_005f0);
-    // /pages/results.jsp(191,2) null
-    _jspx_th_form_005finput_005f0.setDynamicAttribute(null, "type", "text");
-    // /pages/results.jsp(191,2) null
-    _jspx_th_form_005finput_005f0.setDynamicAttribute(null, "name", "city");
-    // /pages/results.jsp(191,2) name = id type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005finput_005f0.setId("city");
-    // /pages/results.jsp(191,2) name = path type = null reqTime = true required = true fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005finput_005f0.setPath("city");
-    // /pages/results.jsp(191,2) null
-    _jspx_th_form_005finput_005f0.setDynamicAttribute(null, "placeholder", "e.g. Bern");
-    // /pages/results.jsp(191,2) name = tabindex type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005finput_005f0.setTabindex("3");
-    int[] _jspx_push_body_count_form_005finput_005f0 = new int[] { 0 };
-    try {
-      int _jspx_eval_form_005finput_005f0 = _jspx_th_form_005finput_005f0.doStartTag();
-      if (_jspx_th_form_005finput_005f0.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-        return true;
-      }
-    } catch (java.lang.Throwable _jspx_exception) {
-      while (_jspx_push_body_count_form_005finput_005f0[0]-- > 0)
-        out = _jspx_page_context.popBody();
-      _jspx_th_form_005finput_005f0.doCatch(_jspx_exception);
-    } finally {
-      _jspx_th_form_005finput_005f0.doFinally();
-      _005fjspx_005ftagPool_005fform_005finput_0026_005ftype_005ftabindex_005fplaceholder_005fpath_005fname_005fid_005fnobody.reuse(_jspx_th_form_005finput_005f0);
-    }
-    return false;
-  }
-
-  private boolean _jspx_meth_form_005finput_005f1(javax.servlet.jsp.tagext.JspTag _jspx_th_form_005fform_005f0, javax.servlet.jsp.PageContext _jspx_page_context, int[] _jspx_push_body_count_form_005fform_005f0)
-          throws java.lang.Throwable {
-    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
-    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
-    //  form:input
-    org.springframework.web.servlet.tags.form.InputTag _jspx_th_form_005finput_005f1 = (org.springframework.web.servlet.tags.form.InputTag) _005fjspx_005ftagPool_005fform_005finput_0026_005ftype_005fstep_005fplaceholder_005fpath_005fid_005fnobody.get(org.springframework.web.servlet.tags.form.InputTag.class);
-    _jspx_th_form_005finput_005f1.setPageContext(_jspx_page_context);
-    _jspx_th_form_005finput_005f1.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_form_005fform_005f0);
-    // /pages/results.jsp(196,2) name = id type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005finput_005f1.setId("radiusInput");
-    // /pages/results.jsp(196,2) null
-    _jspx_th_form_005finput_005f1.setDynamicAttribute(null, "type", "number");
-    // /pages/results.jsp(196,2) name = path type = null reqTime = true required = true fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005finput_005f1.setPath("radius");
-    // /pages/results.jsp(196,2) null
-    _jspx_th_form_005finput_005f1.setDynamicAttribute(null, "placeholder", "e.g. 5");
-    // /pages/results.jsp(196,2) null
-    _jspx_th_form_005finput_005f1.setDynamicAttribute(null, "step", "5");
-    int[] _jspx_push_body_count_form_005finput_005f1 = new int[] { 0 };
-    try {
-      int _jspx_eval_form_005finput_005f1 = _jspx_th_form_005finput_005f1.doStartTag();
-      if (_jspx_th_form_005finput_005f1.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-        return true;
-      }
-    } catch (java.lang.Throwable _jspx_exception) {
-      while (_jspx_push_body_count_form_005finput_005f1[0]-- > 0)
-        out = _jspx_page_context.popBody();
-      _jspx_th_form_005finput_005f1.doCatch(_jspx_exception);
-    } finally {
-      _jspx_th_form_005finput_005f1.doFinally();
-      _005fjspx_005ftagPool_005fform_005finput_0026_005ftype_005fstep_005fplaceholder_005fpath_005fid_005fnobody.reuse(_jspx_th_form_005finput_005f1);
-    }
-    return false;
-  }
-
-  private boolean _jspx_meth_form_005finput_005f2(javax.servlet.jsp.tagext.JspTag _jspx_th_form_005fform_005f0, javax.servlet.jsp.PageContext _jspx_page_context, int[] _jspx_push_body_count_form_005fform_005f0)
-          throws java.lang.Throwable {
-    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
-    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
-    //  form:input
-    org.springframework.web.servlet.tags.form.InputTag _jspx_th_form_005finput_005f2 = (org.springframework.web.servlet.tags.form.InputTag) _005fjspx_005ftagPool_005fform_005finput_0026_005ftype_005fstep_005fplaceholder_005fpath_005fid_005fnobody.get(org.springframework.web.servlet.tags.form.InputTag.class);
-    _jspx_th_form_005finput_005f2.setPageContext(_jspx_page_context);
-    _jspx_th_form_005finput_005f2.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_form_005fform_005f0);
-    // /pages/results.jsp(201,2) name = id type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005finput_005f2.setId("prizeInput");
-    // /pages/results.jsp(201,2) null
-    _jspx_th_form_005finput_005f2.setDynamicAttribute(null, "type", "number");
-    // /pages/results.jsp(201,2) name = path type = null reqTime = true required = true fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005finput_005f2.setPath("prize");
-    // /pages/results.jsp(201,2) null
-    _jspx_th_form_005finput_005f2.setDynamicAttribute(null, "placeholder", "e.g. 5");
-    // /pages/results.jsp(201,2) null
-    _jspx_th_form_005finput_005f2.setDynamicAttribute(null, "step", "50");
-    int[] _jspx_push_body_count_form_005finput_005f2 = new int[] { 0 };
-    try {
-      int _jspx_eval_form_005finput_005f2 = _jspx_th_form_005finput_005f2.doStartTag();
-      if (_jspx_th_form_005finput_005f2.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-        return true;
-      }
-    } catch (java.lang.Throwable _jspx_exception) {
-      while (_jspx_push_body_count_form_005finput_005f2[0]-- > 0)
-        out = _jspx_page_context.popBody();
-      _jspx_th_form_005finput_005f2.doCatch(_jspx_exception);
-    } finally {
-      _jspx_th_form_005finput_005f2.doFinally();
-      _005fjspx_005ftagPool_005fform_005finput_0026_005ftype_005fstep_005fplaceholder_005fpath_005fid_005fnobody.reuse(_jspx_th_form_005finput_005f2);
-    }
-    return false;
-  }
-
-  private boolean _jspx_meth_form_005finput_005f3(javax.servlet.jsp.tagext.JspTag _jspx_th_form_005fform_005f0, javax.servlet.jsp.PageContext _jspx_page_context, int[] _jspx_push_body_count_form_005fform_005f0)
-          throws java.lang.Throwable {
-    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
-    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
-    //  form:input
-    org.springframework.web.servlet.tags.form.InputTag _jspx_th_form_005finput_005f3 = (org.springframework.web.servlet.tags.form.InputTag) _005fjspx_005ftagPool_005fform_005finput_0026_005ftype_005fpath_005fid_005fnobody.get(org.springframework.web.servlet.tags.form.InputTag.class);
-    _jspx_th_form_005finput_005f3.setPageContext(_jspx_page_context);
-    _jspx_th_form_005finput_005f3.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_form_005fform_005f0);
-    // /pages/results.jsp(214,8) null
-    _jspx_th_form_005finput_005f3.setDynamicAttribute(null, "type", "text");
-    // /pages/results.jsp(214,8) name = id type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005finput_005f3.setId("field-earliestMoveInDate");
-    // /pages/results.jsp(214,8) name = path type = null reqTime = true required = true fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005finput_005f3.setPath("earliestMoveInDate");
-    int[] _jspx_push_body_count_form_005finput_005f3 = new int[] { 0 };
-    try {
-      int _jspx_eval_form_005finput_005f3 = _jspx_th_form_005finput_005f3.doStartTag();
-      if (_jspx_th_form_005finput_005f3.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-        return true;
-      }
-    } catch (java.lang.Throwable _jspx_exception) {
-      while (_jspx_push_body_count_form_005finput_005f3[0]-- > 0)
-        out = _jspx_page_context.popBody();
-      _jspx_th_form_005finput_005f3.doCatch(_jspx_exception);
-    } finally {
-      _jspx_th_form_005finput_005f3.doFinally();
-      _005fjspx_005ftagPool_005fform_005finput_0026_005ftype_005fpath_005fid_005fnobody.reuse(_jspx_th_form_005finput_005f3);
-    }
-    return false;
-  }
-
-  private boolean _jspx_meth_form_005finput_005f4(javax.servlet.jsp.tagext.JspTag _jspx_th_form_005fform_005f0, javax.servlet.jsp.PageContext _jspx_page_context, int[] _jspx_push_body_count_form_005fform_005f0)
-          throws java.lang.Throwable {
-    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
-    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
-    //  form:input
-    org.springframework.web.servlet.tags.form.InputTag _jspx_th_form_005finput_005f4 = (org.springframework.web.servlet.tags.form.InputTag) _005fjspx_005ftagPool_005fform_005finput_0026_005ftype_005fpath_005fid_005fnobody.get(org.springframework.web.servlet.tags.form.InputTag.class);
-    _jspx_th_form_005finput_005f4.setPageContext(_jspx_page_context);
-    _jspx_th_form_005finput_005f4.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_form_005fform_005f0);
-    // /pages/results.jsp(216,8) null
-    _jspx_th_form_005finput_005f4.setDynamicAttribute(null, "type", "text");
-    // /pages/results.jsp(216,8) name = id type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005finput_005f4.setId("field-earliestMoveOutDate");
-    // /pages/results.jsp(216,8) name = path type = null reqTime = true required = true fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005finput_005f4.setPath("earliestMoveOutDate");
-    int[] _jspx_push_body_count_form_005finput_005f4 = new int[] { 0 };
-    try {
-      int _jspx_eval_form_005finput_005f4 = _jspx_th_form_005finput_005f4.doStartTag();
-      if (_jspx_th_form_005finput_005f4.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-        return true;
-      }
-    } catch (java.lang.Throwable _jspx_exception) {
-      while (_jspx_push_body_count_form_005finput_005f4[0]-- > 0)
-        out = _jspx_page_context.popBody();
-      _jspx_th_form_005finput_005f4.doCatch(_jspx_exception);
-    } finally {
-      _jspx_th_form_005finput_005f4.doFinally();
-      _005fjspx_005ftagPool_005fform_005finput_0026_005ftype_005fpath_005fid_005fnobody.reuse(_jspx_th_form_005finput_005f4);
-    }
-    return false;
-  }
-
-  private boolean _jspx_meth_form_005finput_005f5(javax.servlet.jsp.tagext.JspTag _jspx_th_form_005fform_005f0, javax.servlet.jsp.PageContext _jspx_page_context, int[] _jspx_push_body_count_form_005fform_005f0)
-          throws java.lang.Throwable {
-    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
-    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
-    //  form:input
-    org.springframework.web.servlet.tags.form.InputTag _jspx_th_form_005finput_005f5 = (org.springframework.web.servlet.tags.form.InputTag) _005fjspx_005ftagPool_005fform_005finput_0026_005ftype_005fpath_005fid_005fnobody.get(org.springframework.web.servlet.tags.form.InputTag.class);
-    _jspx_th_form_005finput_005f5.setPageContext(_jspx_page_context);
-    _jspx_th_form_005finput_005f5.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_form_005fform_005f0);
-    // /pages/results.jsp(224,8) null
-    _jspx_th_form_005finput_005f5.setDynamicAttribute(null, "type", "text");
-    // /pages/results.jsp(224,8) name = id type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005finput_005f5.setId("field-latestMoveInDate");
-    // /pages/results.jsp(224,8) name = path type = null reqTime = true required = true fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005finput_005f5.setPath("latestMoveInDate");
-    int[] _jspx_push_body_count_form_005finput_005f5 = new int[] { 0 };
-    try {
-      int _jspx_eval_form_005finput_005f5 = _jspx_th_form_005finput_005f5.doStartTag();
-      if (_jspx_th_form_005finput_005f5.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-        return true;
-      }
-    } catch (java.lang.Throwable _jspx_exception) {
-      while (_jspx_push_body_count_form_005finput_005f5[0]-- > 0)
-        out = _jspx_page_context.popBody();
-      _jspx_th_form_005finput_005f5.doCatch(_jspx_exception);
-    } finally {
-      _jspx_th_form_005finput_005f5.doFinally();
-      _005fjspx_005ftagPool_005fform_005finput_0026_005ftype_005fpath_005fid_005fnobody.reuse(_jspx_th_form_005finput_005f5);
-    }
-    return false;
-  }
-
-  private boolean _jspx_meth_form_005finput_005f6(javax.servlet.jsp.tagext.JspTag _jspx_th_form_005fform_005f0, javax.servlet.jsp.PageContext _jspx_page_context, int[] _jspx_push_body_count_form_005fform_005f0)
-          throws java.lang.Throwable {
-    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
-    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
-    //  form:input
-    org.springframework.web.servlet.tags.form.InputTag _jspx_th_form_005finput_005f6 = (org.springframework.web.servlet.tags.form.InputTag) _005fjspx_005ftagPool_005fform_005finput_0026_005ftype_005fpath_005fid_005fnobody.get(org.springframework.web.servlet.tags.form.InputTag.class);
-    _jspx_th_form_005finput_005f6.setPageContext(_jspx_page_context);
-    _jspx_th_form_005finput_005f6.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_form_005fform_005f0);
-    // /pages/results.jsp(226,8) null
-    _jspx_th_form_005finput_005f6.setDynamicAttribute(null, "type", "text");
-    // /pages/results.jsp(226,8) name = id type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005finput_005f6.setId("field-latestMoveOutDate");
-    // /pages/results.jsp(226,8) name = path type = null reqTime = true required = true fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005finput_005f6.setPath("latestMoveOutDate");
-    int[] _jspx_push_body_count_form_005finput_005f6 = new int[] { 0 };
-    try {
-      int _jspx_eval_form_005finput_005f6 = _jspx_th_form_005finput_005f6.doStartTag();
-      if (_jspx_th_form_005finput_005f6.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-        return true;
-      }
-    } catch (java.lang.Throwable _jspx_exception) {
-      while (_jspx_push_body_count_form_005finput_005f6[0]-- > 0)
-        out = _jspx_page_context.popBody();
-      _jspx_th_form_005finput_005f6.doCatch(_jspx_exception);
-    } finally {
-      _jspx_th_form_005finput_005f6.doFinally();
-      _005fjspx_005ftagPool_005fform_005finput_0026_005ftype_005fpath_005fid_005fnobody.reuse(_jspx_th_form_005finput_005f6);
-    }
-    return false;
-  }
-
-  private boolean _jspx_meth_form_005fcheckbox_005f6(javax.servlet.jsp.tagext.JspTag _jspx_th_form_005fform_005f0, javax.servlet.jsp.PageContext _jspx_page_context, int[] _jspx_push_body_count_form_005fform_005f0)
-          throws java.lang.Throwable {
-    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
-    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
-    //  form:checkbox
-    org.springframework.web.servlet.tags.form.CheckboxTag _jspx_th_form_005fcheckbox_005f6 = (org.springframework.web.servlet.tags.form.CheckboxTag) _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fvalue_005fpath_005fid_005fnobody.get(org.springframework.web.servlet.tags.form.CheckboxTag.class);
-    _jspx_th_form_005fcheckbox_005f6.setPageContext(_jspx_page_context);
-    _jspx_th_form_005fcheckbox_005f6.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_form_005fform_005f0);
-    // /pages/results.jsp(230,8) name = id type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005fcheckbox_005f6.setId("field-smoker");
-    // /pages/results.jsp(230,8) name = path type = null reqTime = true required = true fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005fcheckbox_005f6.setPath("smokers");
-    // /pages/results.jsp(230,8) name = value type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005fcheckbox_005f6.setValue("1");
-    int[] _jspx_push_body_count_form_005fcheckbox_005f6 = new int[] { 0 };
-    try {
-      int _jspx_eval_form_005fcheckbox_005f6 = _jspx_th_form_005fcheckbox_005f6.doStartTag();
-      if (_jspx_th_form_005fcheckbox_005f6.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-        return true;
-      }
-    } catch (java.lang.Throwable _jspx_exception) {
-      while (_jspx_push_body_count_form_005fcheckbox_005f6[0]-- > 0)
-        out = _jspx_page_context.popBody();
-      _jspx_th_form_005fcheckbox_005f6.doCatch(_jspx_exception);
-    } finally {
-      _jspx_th_form_005fcheckbox_005f6.doFinally();
-      _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fvalue_005fpath_005fid_005fnobody.reuse(_jspx_th_form_005fcheckbox_005f6);
-    }
-    return false;
-  }
-
-  private boolean _jspx_meth_form_005fcheckbox_005f7(javax.servlet.jsp.tagext.JspTag _jspx_th_form_005fform_005f0, javax.servlet.jsp.PageContext _jspx_page_context, int[] _jspx_push_body_count_form_005fform_005f0)
-          throws java.lang.Throwable {
-    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
-    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
-    //  form:checkbox
-    org.springframework.web.servlet.tags.form.CheckboxTag _jspx_th_form_005fcheckbox_005f7 = (org.springframework.web.servlet.tags.form.CheckboxTag) _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fvalue_005fpath_005fid_005fnobody.get(org.springframework.web.servlet.tags.form.CheckboxTag.class);
-    _jspx_th_form_005fcheckbox_005f7.setPageContext(_jspx_page_context);
-    _jspx_th_form_005fcheckbox_005f7.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_form_005fform_005f0);
-    // /pages/results.jsp(232,8) name = id type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005fcheckbox_005f7.setId("field-animals");
-    // /pages/results.jsp(232,8) name = path type = null reqTime = true required = true fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005fcheckbox_005f7.setPath("animals");
-    // /pages/results.jsp(232,8) name = value type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005fcheckbox_005f7.setValue("1");
-    int[] _jspx_push_body_count_form_005fcheckbox_005f7 = new int[] { 0 };
-    try {
-      int _jspx_eval_form_005fcheckbox_005f7 = _jspx_th_form_005fcheckbox_005f7.doStartTag();
-      if (_jspx_th_form_005fcheckbox_005f7.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-        return true;
-      }
-    } catch (java.lang.Throwable _jspx_exception) {
-      while (_jspx_push_body_count_form_005fcheckbox_005f7[0]-- > 0)
-        out = _jspx_page_context.popBody();
-      _jspx_th_form_005fcheckbox_005f7.doCatch(_jspx_exception);
-    } finally {
-      _jspx_th_form_005fcheckbox_005f7.doFinally();
-      _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fvalue_005fpath_005fid_005fnobody.reuse(_jspx_th_form_005fcheckbox_005f7);
-    }
-    return false;
-  }
-
-  private boolean _jspx_meth_form_005fcheckbox_005f8(javax.servlet.jsp.tagext.JspTag _jspx_th_form_005fform_005f0, javax.servlet.jsp.PageContext _jspx_page_context, int[] _jspx_push_body_count_form_005fform_005f0)
-          throws java.lang.Throwable {
-    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
-    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
-    //  form:checkbox
-    org.springframework.web.servlet.tags.form.CheckboxTag _jspx_th_form_005fcheckbox_005f8 = (org.springframework.web.servlet.tags.form.CheckboxTag) _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fvalue_005fpath_005fid_005fnobody.get(org.springframework.web.servlet.tags.form.CheckboxTag.class);
-    _jspx_th_form_005fcheckbox_005f8.setPageContext(_jspx_page_context);
-    _jspx_th_form_005fcheckbox_005f8.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_form_005fform_005f0);
-    // /pages/results.jsp(236,8) name = id type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005fcheckbox_005f8.setId("field-garden");
-    // /pages/results.jsp(236,8) name = path type = null reqTime = true required = true fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005fcheckbox_005f8.setPath("garden");
-    // /pages/results.jsp(236,8) name = value type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005fcheckbox_005f8.setValue("1");
-    int[] _jspx_push_body_count_form_005fcheckbox_005f8 = new int[] { 0 };
-    try {
-      int _jspx_eval_form_005fcheckbox_005f8 = _jspx_th_form_005fcheckbox_005f8.doStartTag();
-      if (_jspx_th_form_005fcheckbox_005f8.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-        return true;
-      }
-    } catch (java.lang.Throwable _jspx_exception) {
-      while (_jspx_push_body_count_form_005fcheckbox_005f8[0]-- > 0)
-        out = _jspx_page_context.popBody();
-      _jspx_th_form_005fcheckbox_005f8.doCatch(_jspx_exception);
-    } finally {
-      _jspx_th_form_005fcheckbox_005f8.doFinally();
-      _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fvalue_005fpath_005fid_005fnobody.reuse(_jspx_th_form_005fcheckbox_005f8);
-    }
-    return false;
-  }
-
-  private boolean _jspx_meth_form_005fcheckbox_005f9(javax.servlet.jsp.tagext.JspTag _jspx_th_form_005fform_005f0, javax.servlet.jsp.PageContext _jspx_page_context, int[] _jspx_push_body_count_form_005fform_005f0)
-          throws java.lang.Throwable {
-    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
-    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
-    //  form:checkbox
-    org.springframework.web.servlet.tags.form.CheckboxTag _jspx_th_form_005fcheckbox_005f9 = (org.springframework.web.servlet.tags.form.CheckboxTag) _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fvalue_005fpath_005fid_005fnobody.get(org.springframework.web.servlet.tags.form.CheckboxTag.class);
-    _jspx_th_form_005fcheckbox_005f9.setPageContext(_jspx_page_context);
-    _jspx_th_form_005fcheckbox_005f9.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_form_005fform_005f0);
-    // /pages/results.jsp(238,8) name = id type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005fcheckbox_005f9.setId("field-balcony");
-    // /pages/results.jsp(238,8) name = path type = null reqTime = true required = true fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005fcheckbox_005f9.setPath("balcony");
-    // /pages/results.jsp(238,8) name = value type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005fcheckbox_005f9.setValue("1");
-    int[] _jspx_push_body_count_form_005fcheckbox_005f9 = new int[] { 0 };
-    try {
-      int _jspx_eval_form_005fcheckbox_005f9 = _jspx_th_form_005fcheckbox_005f9.doStartTag();
-      if (_jspx_th_form_005fcheckbox_005f9.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-        return true;
-      }
-    } catch (java.lang.Throwable _jspx_exception) {
-      while (_jspx_push_body_count_form_005fcheckbox_005f9[0]-- > 0)
-        out = _jspx_page_context.popBody();
-      _jspx_th_form_005fcheckbox_005f9.doCatch(_jspx_exception);
-    } finally {
-      _jspx_th_form_005fcheckbox_005f9.doFinally();
-      _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fvalue_005fpath_005fid_005fnobody.reuse(_jspx_th_form_005fcheckbox_005f9);
-    }
-    return false;
-  }
-
-  private boolean _jspx_meth_form_005fcheckbox_005f10(javax.servlet.jsp.tagext.JspTag _jspx_th_form_005fform_005f0, javax.servlet.jsp.PageContext _jspx_page_context, int[] _jspx_push_body_count_form_005fform_005f0)
-          throws java.lang.Throwable {
-    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
-    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
-    //  form:checkbox
-    org.springframework.web.servlet.tags.form.CheckboxTag _jspx_th_form_005fcheckbox_005f10 = (org.springframework.web.servlet.tags.form.CheckboxTag) _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fvalue_005fpath_005fid_005fnobody.get(org.springframework.web.servlet.tags.form.CheckboxTag.class);
-    _jspx_th_form_005fcheckbox_005f10.setPageContext(_jspx_page_context);
-    _jspx_th_form_005fcheckbox_005f10.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_form_005fform_005f0);
-    // /pages/results.jsp(242,8) name = id type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005fcheckbox_005f10.setId("field-cellar");
-    // /pages/results.jsp(242,8) name = path type = null reqTime = true required = true fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005fcheckbox_005f10.setPath("cellar");
-    // /pages/results.jsp(242,8) name = value type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005fcheckbox_005f10.setValue("1");
-    int[] _jspx_push_body_count_form_005fcheckbox_005f10 = new int[] { 0 };
-    try {
-      int _jspx_eval_form_005fcheckbox_005f10 = _jspx_th_form_005fcheckbox_005f10.doStartTag();
-      if (_jspx_th_form_005fcheckbox_005f10.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-        return true;
-      }
-    } catch (java.lang.Throwable _jspx_exception) {
-      while (_jspx_push_body_count_form_005fcheckbox_005f10[0]-- > 0)
-        out = _jspx_page_context.popBody();
-      _jspx_th_form_005fcheckbox_005f10.doCatch(_jspx_exception);
-    } finally {
-      _jspx_th_form_005fcheckbox_005f10.doFinally();
-      _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fvalue_005fpath_005fid_005fnobody.reuse(_jspx_th_form_005fcheckbox_005f10);
-    }
-    return false;
-  }
-
-  private boolean _jspx_meth_form_005fcheckbox_005f11(javax.servlet.jsp.tagext.JspTag _jspx_th_form_005fform_005f0, javax.servlet.jsp.PageContext _jspx_page_context, int[] _jspx_push_body_count_form_005fform_005f0)
-          throws java.lang.Throwable {
-    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
-    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
-    //  form:checkbox
-    org.springframework.web.servlet.tags.form.CheckboxTag _jspx_th_form_005fcheckbox_005f11 = (org.springframework.web.servlet.tags.form.CheckboxTag) _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fvalue_005fpath_005fid_005fnobody.get(org.springframework.web.servlet.tags.form.CheckboxTag.class);
-    _jspx_th_form_005fcheckbox_005f11.setPageContext(_jspx_page_context);
-    _jspx_th_form_005fcheckbox_005f11.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_form_005fform_005f0);
-    // /pages/results.jsp(244,8) name = id type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005fcheckbox_005f11.setId("field-furnished");
-    // /pages/results.jsp(244,8) name = path type = null reqTime = true required = true fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005fcheckbox_005f11.setPath("furnished");
-    // /pages/results.jsp(244,8) name = value type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005fcheckbox_005f11.setValue("1");
-    int[] _jspx_push_body_count_form_005fcheckbox_005f11 = new int[] { 0 };
-    try {
-      int _jspx_eval_form_005fcheckbox_005f11 = _jspx_th_form_005fcheckbox_005f11.doStartTag();
-      if (_jspx_th_form_005fcheckbox_005f11.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-        return true;
-      }
-    } catch (java.lang.Throwable _jspx_exception) {
-      while (_jspx_push_body_count_form_005fcheckbox_005f11[0]-- > 0)
-        out = _jspx_page_context.popBody();
-      _jspx_th_form_005fcheckbox_005f11.doCatch(_jspx_exception);
-    } finally {
-      _jspx_th_form_005fcheckbox_005f11.doFinally();
-      _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fvalue_005fpath_005fid_005fnobody.reuse(_jspx_th_form_005fcheckbox_005f11);
-    }
-    return false;
-  }
-
-  private boolean _jspx_meth_form_005fcheckbox_005f12(javax.servlet.jsp.tagext.JspTag _jspx_th_form_005fform_005f0, javax.servlet.jsp.PageContext _jspx_page_context, int[] _jspx_push_body_count_form_005fform_005f0)
-          throws java.lang.Throwable {
-    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
-    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
-    //  form:checkbox
-    org.springframework.web.servlet.tags.form.CheckboxTag _jspx_th_form_005fcheckbox_005f12 = (org.springframework.web.servlet.tags.form.CheckboxTag) _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fvalue_005fpath_005fid_005fnobody.get(org.springframework.web.servlet.tags.form.CheckboxTag.class);
-    _jspx_th_form_005fcheckbox_005f12.setPageContext(_jspx_page_context);
-    _jspx_th_form_005fcheckbox_005f12.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_form_005fform_005f0);
-    // /pages/results.jsp(248,8) name = id type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005fcheckbox_005f12.setId("field-cable");
-    // /pages/results.jsp(248,8) name = path type = null reqTime = true required = true fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005fcheckbox_005f12.setPath("cable");
-    // /pages/results.jsp(248,8) name = value type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005fcheckbox_005f12.setValue("1");
-    int[] _jspx_push_body_count_form_005fcheckbox_005f12 = new int[] { 0 };
-    try {
-      int _jspx_eval_form_005fcheckbox_005f12 = _jspx_th_form_005fcheckbox_005f12.doStartTag();
-      if (_jspx_th_form_005fcheckbox_005f12.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-        return true;
-      }
-    } catch (java.lang.Throwable _jspx_exception) {
-      while (_jspx_push_body_count_form_005fcheckbox_005f12[0]-- > 0)
-        out = _jspx_page_context.popBody();
-      _jspx_th_form_005fcheckbox_005f12.doCatch(_jspx_exception);
-    } finally {
-      _jspx_th_form_005fcheckbox_005f12.doFinally();
-      _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fvalue_005fpath_005fid_005fnobody.reuse(_jspx_th_form_005fcheckbox_005f12);
-    }
-    return false;
-  }
-
-  private boolean _jspx_meth_form_005fcheckbox_005f13(javax.servlet.jsp.tagext.JspTag _jspx_th_form_005fform_005f0, javax.servlet.jsp.PageContext _jspx_page_context, int[] _jspx_push_body_count_form_005fform_005f0)
-          throws java.lang.Throwable {
-    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
-    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
-    //  form:checkbox
-    org.springframework.web.servlet.tags.form.CheckboxTag _jspx_th_form_005fcheckbox_005f13 = (org.springframework.web.servlet.tags.form.CheckboxTag) _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fvalue_005fpath_005fid_005fnobody.get(org.springframework.web.servlet.tags.form.CheckboxTag.class);
-    _jspx_th_form_005fcheckbox_005f13.setPageContext(_jspx_page_context);
-    _jspx_th_form_005fcheckbox_005f13.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_form_005fform_005f0);
-    // /pages/results.jsp(250,8) name = id type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005fcheckbox_005f13.setId("field-garage");
-    // /pages/results.jsp(250,8) name = path type = null reqTime = true required = true fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005fcheckbox_005f13.setPath("garage");
-    // /pages/results.jsp(250,8) name = value type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005fcheckbox_005f13.setValue("1");
-    int[] _jspx_push_body_count_form_005fcheckbox_005f13 = new int[] { 0 };
-    try {
-      int _jspx_eval_form_005fcheckbox_005f13 = _jspx_th_form_005fcheckbox_005f13.doStartTag();
-      if (_jspx_th_form_005fcheckbox_005f13.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-        return true;
-      }
-    } catch (java.lang.Throwable _jspx_exception) {
-      while (_jspx_push_body_count_form_005fcheckbox_005f13[0]-- > 0)
-        out = _jspx_page_context.popBody();
-      _jspx_th_form_005fcheckbox_005f13.doCatch(_jspx_exception);
-    } finally {
-      _jspx_th_form_005fcheckbox_005f13.doFinally();
-      _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fvalue_005fpath_005fid_005fnobody.reuse(_jspx_th_form_005fcheckbox_005f13);
-    }
-    return false;
-  }
-
-  private boolean _jspx_meth_form_005fcheckbox_005f14(javax.servlet.jsp.tagext.JspTag _jspx_th_form_005fform_005f0, javax.servlet.jsp.PageContext _jspx_page_context, int[] _jspx_push_body_count_form_005fform_005f0)
-          throws java.lang.Throwable {
-    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
-    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
-    //  form:checkbox
-    org.springframework.web.servlet.tags.form.CheckboxTag _jspx_th_form_005fcheckbox_005f14 = (org.springframework.web.servlet.tags.form.CheckboxTag) _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fvalue_005fpath_005fid_005fnobody.get(org.springframework.web.servlet.tags.form.CheckboxTag.class);
-    _jspx_th_form_005fcheckbox_005f14.setPageContext(_jspx_page_context);
-    _jspx_th_form_005fcheckbox_005f14.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_form_005fform_005f0);
-    // /pages/results.jsp(254,8) name = id type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005fcheckbox_005f14.setId("field-internet");
-    // /pages/results.jsp(254,8) name = path type = null reqTime = true required = true fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005fcheckbox_005f14.setPath("internet");
-    // /pages/results.jsp(254,8) name = value type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005fcheckbox_005f14.setValue("1");
-    int[] _jspx_push_body_count_form_005fcheckbox_005f14 = new int[] { 0 };
-    try {
-      int _jspx_eval_form_005fcheckbox_005f14 = _jspx_th_form_005fcheckbox_005f14.doStartTag();
-      if (_jspx_th_form_005fcheckbox_005f14.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-        return true;
-      }
-    } catch (java.lang.Throwable _jspx_exception) {
-      while (_jspx_push_body_count_form_005fcheckbox_005f14[0]-- > 0)
-        out = _jspx_page_context.popBody();
-      _jspx_th_form_005fcheckbox_005f14.doCatch(_jspx_exception);
-    } finally {
-      _jspx_th_form_005fcheckbox_005f14.doFinally();
-      _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fvalue_005fpath_005fid_005fnobody.reuse(_jspx_th_form_005fcheckbox_005f14);
-    }
-    return false;
-  }
-
-  private boolean _jspx_meth_c_005fimport_005f2(javax.servlet.jsp.PageContext _jspx_page_context)
-          throws java.lang.Throwable {
-    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
-    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
-    //  c:import
-    org.apache.taglibs.standard.tag.rt.core.ImportTag _jspx_th_c_005fimport_005f2 = (org.apache.taglibs.standard.tag.rt.core.ImportTag) _005fjspx_005ftagPool_005fc_005fimport_0026_005furl_005fnobody.get(org.apache.taglibs.standard.tag.rt.core.ImportTag.class);
-    _jspx_th_c_005fimport_005f2.setPageContext(_jspx_page_context);
-    _jspx_th_c_005fimport_005f2.setParent(null);
-    // /pages/results.jsp(264,0) name = url type = null reqTime = true required = true fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_c_005fimport_005f2.setUrl("template/footer.jsp");
-    int[] _jspx_push_body_count_c_005fimport_005f2 = new int[] { 0 };
-    try {
-      int _jspx_eval_c_005fimport_005f2 = _jspx_th_c_005fimport_005f2.doStartTag();
-      if (_jspx_th_c_005fimport_005f2.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-        return true;
-      }
-    } catch (java.lang.Throwable _jspx_exception) {
-      while (_jspx_push_body_count_c_005fimport_005f2[0]-- > 0)
-        out = _jspx_page_context.popBody();
-      _jspx_th_c_005fimport_005f2.doCatch(_jspx_exception);
-    } finally {
-      _jspx_th_c_005fimport_005f2.doFinally();
-      _005fjspx_005ftagPool_005fc_005fimport_0026_005furl_005fnobody.reuse(_jspx_th_c_005fimport_005f2);
-    }
-    return false;
-  }
-}
