--- baseProject/temp/jsp/org/apache/jsp/pages/searchAd_jsp.java
@@ -1,752 +1,0 @@
-/*
- * Generated by the Jasper component of Apache Tomcat
- * Version: jetty/9.2.3.v20140905
- * Generated at: 2016-10-04 18:23:05 UTC
- * Note: The last modified time of this file was set to
- *       the last modified time of the source file after
- *       generation to assist with modification tracking.
- */
-package org.apache.jsp.pages;
-
-import javax.servlet.*;
-import javax.servlet.http.*;
-import javax.servlet.jsp.*;
-
-public final class searchAd_jsp extends org.apache.jasper.runtime.HttpJspBase
-    implements org.apache.jasper.runtime.JspSourceDependent {
-
-  private static final javax.servlet.jsp.JspFactory _jspxFactory =
-          javax.servlet.jsp.JspFactory.getDefaultFactory();
-
-  private static java.util.Map<java.lang.String,java.lang.Long> _jspx_dependants;
-
-  static {
-    _jspx_dependants = new java.util.HashMap<java.lang.String,java.lang.Long>(5);
-    _jspx_dependants.put("file:/C:/Users/Julian%20Schuhmacher/.m2/repository/org/springframework/spring-webmvc/4.1.1.RELEASE/spring-webmvc-4.1.1.RELEASE.jar", Long.valueOf(1475180010191L));
-    _jspx_dependants.put("jar:file:/C:/Users/Julian%20Schuhmacher/.m2/repository/org/apache/taglibs/taglibs-standard-impl/1.2.1/taglibs-standard-impl-1.2.1.jar!/META-INF/c.tld", Long.valueOf(1384364862000L));
-    _jspx_dependants.put("file:/C:/Users/Julian%20Schuhmacher/.m2/repository/org/apache/taglibs/taglibs-standard-impl/1.2.1/taglibs-standard-impl-1.2.1.jar", Long.valueOf(1475217457864L));
-    _jspx_dependants.put("jar:file:/C:/Users/Julian%20Schuhmacher/.m2/repository/org/springframework/spring-webmvc/4.1.1.RELEASE/spring-webmvc-4.1.1.RELEASE.jar!/META-INF/spring-form.tld", Long.valueOf(1412119640000L));
-    _jspx_dependants.put("jar:file:/C:/Users/Julian%20Schuhmacher/.m2/repository/org/springframework/spring-webmvc/4.1.1.RELEASE/spring-webmvc-4.1.1.RELEASE.jar!/META-INF/spring.tld", Long.valueOf(1412119640000L));
-  }
-
-  private org.apache.jasper.runtime.TagHandlerPool _005fjspx_005ftagPool_005fc_005fimport_0026_005furl_005fnobody;
-  private org.apache.jasper.runtime.TagHandlerPool _005fjspx_005ftagPool_005fform_005fform_0026_005fmodelAttribute_005fmethod_005fid_005fautocomplete_005faction;
-  private org.apache.jasper.runtime.TagHandlerPool _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fpath_005fname_005fid_005fnobody;
-  private org.apache.jasper.runtime.TagHandlerPool _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fstyle_005fpath_005fname_005fid_005fnobody;
-  private org.apache.jasper.runtime.TagHandlerPool _005fjspx_005ftagPool_005fform_005ferrors_0026_005fpath_005fcssClass_005fnobody;
-  private org.apache.jasper.runtime.TagHandlerPool _005fjspx_005ftagPool_005fform_005finput_0026_005ftype_005ftabindex_005fplaceholder_005fpath_005fname_005fid_005fnobody;
-  private org.apache.jasper.runtime.TagHandlerPool _005fjspx_005ftagPool_005fform_005finput_0026_005ftype_005fstep_005fplaceholder_005fpath_005fid_005fnobody;
-
-  private javax.el.ExpressionFactory _el_expressionfactory;
-  private org.apache.tomcat.InstanceManager _jsp_instancemanager;
-
-  public java.util.Map<java.lang.String,java.lang.Long> getDependants() {
-    return _jspx_dependants;
-  }
-
-  public void _jspInit() {
-    _005fjspx_005ftagPool_005fc_005fimport_0026_005furl_005fnobody = org.apache.jasper.runtime.TagHandlerPool.getTagHandlerPool(getServletConfig());
-    _005fjspx_005ftagPool_005fform_005fform_0026_005fmodelAttribute_005fmethod_005fid_005fautocomplete_005faction = org.apache.jasper.runtime.TagHandlerPool.getTagHandlerPool(getServletConfig());
-    _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fpath_005fname_005fid_005fnobody = org.apache.jasper.runtime.TagHandlerPool.getTagHandlerPool(getServletConfig());
-    _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fstyle_005fpath_005fname_005fid_005fnobody = org.apache.jasper.runtime.TagHandlerPool.getTagHandlerPool(getServletConfig());
-    _005fjspx_005ftagPool_005fform_005ferrors_0026_005fpath_005fcssClass_005fnobody = org.apache.jasper.runtime.TagHandlerPool.getTagHandlerPool(getServletConfig());
-    _005fjspx_005ftagPool_005fform_005finput_0026_005ftype_005ftabindex_005fplaceholder_005fpath_005fname_005fid_005fnobody = org.apache.jasper.runtime.TagHandlerPool.getTagHandlerPool(getServletConfig());
-    _005fjspx_005ftagPool_005fform_005finput_0026_005ftype_005fstep_005fplaceholder_005fpath_005fid_005fnobody = org.apache.jasper.runtime.TagHandlerPool.getTagHandlerPool(getServletConfig());
-    _el_expressionfactory = _jspxFactory.getJspApplicationContext(getServletConfig().getServletContext()).getExpressionFactory();
-    _jsp_instancemanager = org.apache.jasper.runtime.InstanceManagerFactory.getInstanceManager(getServletConfig());
-  }
-
-  public void _jspDestroy() {
-    _005fjspx_005ftagPool_005fc_005fimport_0026_005furl_005fnobody.release();
-    _005fjspx_005ftagPool_005fform_005fform_0026_005fmodelAttribute_005fmethod_005fid_005fautocomplete_005faction.release();
-    _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fpath_005fname_005fid_005fnobody.release();
-    _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fstyle_005fpath_005fname_005fid_005fnobody.release();
-    _005fjspx_005ftagPool_005fform_005ferrors_0026_005fpath_005fcssClass_005fnobody.release();
-    _005fjspx_005ftagPool_005fform_005finput_0026_005ftype_005ftabindex_005fplaceholder_005fpath_005fname_005fid_005fnobody.release();
-    _005fjspx_005ftagPool_005fform_005finput_0026_005ftype_005fstep_005fplaceholder_005fpath_005fid_005fnobody.release();
-  }
-
-  public void _jspService(final javax.servlet.http.HttpServletRequest request, final javax.servlet.http.HttpServletResponse response)
-        throws java.io.IOException, javax.servlet.ServletException {
-
-    final javax.servlet.jsp.PageContext pageContext;
-    javax.servlet.http.HttpSession session = null;
-    final javax.servlet.ServletContext application;
-    final javax.servlet.ServletConfig config;
-    javax.servlet.jsp.JspWriter out = null;
-    final java.lang.Object page = this;
-    javax.servlet.jsp.JspWriter _jspx_out = null;
-    javax.servlet.jsp.PageContext _jspx_page_context = null;
-
-
-    try {
-      response.setContentType("text/html;charset=utf-8");
-      pageContext = _jspxFactory.getPageContext(this, request, response,
-      			null, true, 8192, true);
-      _jspx_page_context = pageContext;
-      application = pageContext.getServletContext();
-      config = pageContext.getServletConfig();
-      session = pageContext.getSession();
-      out = pageContext.getOut();
-      _jspx_out = out;
-
-      out.write("\r\n");
-      out.write("\r\n");
-      out.write("\r\n");
-      out.write("\r\n");
-      out.write("\r\n");
-      if (_jspx_meth_c_005fimport_005f0(_jspx_page_context))
-        return;
-      out.write("\r\n");
-      out.write("\r\n");
-      out.write("<pre><a href=\"/\">Home</a>   &gt;   Search</pre>\r\n");
-      out.write("\r\n");
-      out.write("<script>\r\n");
-      out.write("\t$(document).ready(function() {\r\n");
-      out.write("\t\t$(\"#city\").autocomplete({\r\n");
-      out.write("\t\t\tminLength : 2\r\n");
-      out.write("\t\t});\r\n");
-      out.write("\t\t$(\"#city\").autocomplete({\r\n");
-      out.write("\t\t\tsource : ");
-      if (_jspx_meth_c_005fimport_005f1(_jspx_page_context))
-        return;
-      out.write("\r\n");
-      out.write("\t\t});\r\n");
-      out.write("\t\t$(\"#city\").autocomplete(\"option\", {\r\n");
-      out.write("\t\t\tenabled : true,\r\n");
-      out.write("\t\t\tautoFocus : true\r\n");
-      out.write("\t\t});\r\n");
-      out.write("\t\t\r\n");
-      out.write("\t\tvar price = document.getElementById('prizeInput');\r\n");
-      out.write("\t\tvar radius = document.getElementById('radiusInput');\r\n");
-      out.write("\t\t\r\n");
-      out.write("\t\tif(price.value == null || price.value == \"\" || price.value == \"0\")\r\n");
-      out.write("\t\t\tprice.value = \"500\";\r\n");
-      out.write("\t\tif(radius.value == null || radius.value == \"\" || radius.value == \"0\")\r\n");
-      out.write("\t\t\tradius.value = \"5\";\r\n");
-      out.write("\t});\r\n");
-      out.write("</script>\r\n");
-      out.write("\r\n");
-      out.write("\r\n");
-      out.write("<script>\r\n");
-      out.write("function validateType(form)\r\n");
-      out.write("{\r\n");
-      out.write("\tvar room = document.getElementById('room');\r\n");
-      out.write("\tvar studio = document.getElementById('studio');\r\n");
-      out.write("\tvar neither = document.getElementById('neither');\r\n");
-      out.write("\tvar both = document.getElementById('both');\r\n");
-      out.write("\tvar type = document.getElementById('type');\r\n");
-      out.write("\tvar filtered = document.getElementById('filtered');\r\n");
-      out.write("\t\r\n");
-      out.write("\tif(room.checked && studio.checked) {\r\n");
-      out.write("\t\tboth.checked = true;\r\n");
-      out.write("\t\tneither.checked = false;\r\n");
-      out.write("\t}\r\n");
-      out.write("\telse if(!room.checked && !studio.checked) {\r\n");
-      out.write("\t\tboth.checked = false;\r\n");
-      out.write("\t\tneither.checked = true;\r\n");
-      out.write("\t}\r\n");
-      out.write("\telse {\r\n");
-      out.write("\t\tboth.checked = false;\r\n");
-      out.write("\t\tneither.checked = false;\r\n");
-      out.write("\t\ttype.checked = studio.checked;\r\n");
-      out.write("\t}\r\n");
-      out.write("\tfiltered.checked = false;\r\n");
-      out.write("}\r\n");
-      out.write("</script>\r\n");
-      out.write("\r\n");
-      out.write("<h1>Search for an ad</h1>\r\n");
-      out.write("<hr />\r\n");
-      out.write("\r\n");
-      //  form:form
-      org.springframework.web.servlet.tags.form.FormTag _jspx_th_form_005fform_005f0 = (org.springframework.web.servlet.tags.form.FormTag) _005fjspx_005ftagPool_005fform_005fform_0026_005fmodelAttribute_005fmethod_005fid_005fautocomplete_005faction.get(org.springframework.web.servlet.tags.form.FormTag.class);
-      _jspx_th_form_005fform_005f0.setPageContext(_jspx_page_context);
-      _jspx_th_form_005fform_005f0.setParent(null);
-      // /pages/searchAd.jsp(65,0) name = method type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-      _jspx_th_form_005fform_005f0.setMethod("post");
-      // /pages/searchAd.jsp(65,0) name = modelAttribute type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-      _jspx_th_form_005fform_005f0.setModelAttribute("searchForm");
-      // /pages/searchAd.jsp(65,0) name = action type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-      _jspx_th_form_005fform_005f0.setAction("/results");
-      // /pages/searchAd.jsp(65,0) name = id type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-      _jspx_th_form_005fform_005f0.setId("searchForm");
-      // /pages/searchAd.jsp(65,0) name = autocomplete type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-      _jspx_th_form_005fform_005f0.setAutocomplete("off");
-      int[] _jspx_push_body_count_form_005fform_005f0 = new int[] { 0 };
-      try {
-        int _jspx_eval_form_005fform_005f0 = _jspx_th_form_005fform_005f0.doStartTag();
-        if (_jspx_eval_form_005fform_005f0 != javax.servlet.jsp.tagext.Tag.SKIP_BODY) {
-          do {
-            out.write("\r\n");
-            out.write("\r\n");
-            out.write("\t<fieldset>\r\n");
-            out.write("\t\t");
-            if (_jspx_meth_form_005fcheckbox_005f0(_jspx_th_form_005fform_005f0, _jspx_page_context, _jspx_push_body_count_form_005fform_005f0))
-              return;
-            out.write("<label>Room</label>\r\n");
-            out.write("\t\t");
-            if (_jspx_meth_form_005fcheckbox_005f1(_jspx_th_form_005fform_005f0, _jspx_page_context, _jspx_push_body_count_form_005fform_005f0))
-              return;
-            out.write("<label>Studio</label>\r\n");
-            out.write("\t\t\r\n");
-            out.write("\t\t");
-            if (_jspx_meth_form_005fcheckbox_005f2(_jspx_th_form_005fform_005f0, _jspx_page_context, _jspx_push_body_count_form_005fform_005f0))
-              return;
-            out.write("\r\n");
-            out.write("\t\t");
-            if (_jspx_meth_form_005fcheckbox_005f3(_jspx_th_form_005fform_005f0, _jspx_page_context, _jspx_push_body_count_form_005fform_005f0))
-              return;
-            out.write("\r\n");
-            out.write("\t\t");
-            if (_jspx_meth_form_005fcheckbox_005f4(_jspx_th_form_005fform_005f0, _jspx_page_context, _jspx_push_body_count_form_005fform_005f0))
-              return;
-            out.write("\r\n");
-            out.write("\t\t");
-            if (_jspx_meth_form_005fcheckbox_005f5(_jspx_th_form_005fform_005f0, _jspx_page_context, _jspx_push_body_count_form_005fform_005f0))
-              return;
-            out.write("\r\n");
-            out.write("\t\t");
-            //  form:errors
-            org.springframework.web.servlet.tags.form.ErrorsTag _jspx_th_form_005ferrors_005f0 = (org.springframework.web.servlet.tags.form.ErrorsTag) _005fjspx_005ftagPool_005fform_005ferrors_0026_005fpath_005fcssClass_005fnobody.get(org.springframework.web.servlet.tags.form.ErrorsTag.class);
-            _jspx_th_form_005ferrors_005f0.setPageContext(_jspx_page_context);
-            _jspx_th_form_005ferrors_005f0.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_form_005fform_005f0);
-            // /pages/searchAd.jsp(76,2) name = path type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-            _jspx_th_form_005ferrors_005f0.setPath("noRoomNoStudio");
-            // /pages/searchAd.jsp(76,2) name = cssClass type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-            _jspx_th_form_005ferrors_005f0.setCssClass("validationErrorText");
-            int[] _jspx_push_body_count_form_005ferrors_005f0 = new int[] { 0 };
-            try {
-              int _jspx_eval_form_005ferrors_005f0 = _jspx_th_form_005ferrors_005f0.doStartTag();
-              if (_jspx_th_form_005ferrors_005f0.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-                return;
-              }
-            } catch (java.lang.Throwable _jspx_exception) {
-              while (_jspx_push_body_count_form_005ferrors_005f0[0]-- > 0)
-                out = _jspx_page_context.popBody();
-              _jspx_th_form_005ferrors_005f0.doCatch(_jspx_exception);
-            } finally {
-              _jspx_th_form_005ferrors_005f0.doFinally();
-              _005fjspx_005ftagPool_005fform_005ferrors_0026_005fpath_005fcssClass_005fnobody.reuse(_jspx_th_form_005ferrors_005f0);
-            }
-            out.write("\r\n");
-            out.write("\t\t\r\n");
-            out.write("\t\t<br />\r\n");
-            out.write("\t\t\r\n");
-            out.write("\t\t<label for=\"city\">City / zip code:</label>\r\n");
-            out.write("\t\t");
-            if (_jspx_meth_form_005finput_005f0(_jspx_th_form_005fform_005f0, _jspx_page_context, _jspx_push_body_count_form_005fform_005f0))
-              return;
-            out.write("\r\n");
-            out.write("\t\t");
-            //  form:errors
-            org.springframework.web.servlet.tags.form.ErrorsTag _jspx_th_form_005ferrors_005f1 = (org.springframework.web.servlet.tags.form.ErrorsTag) _005fjspx_005ftagPool_005fform_005ferrors_0026_005fpath_005fcssClass_005fnobody.get(org.springframework.web.servlet.tags.form.ErrorsTag.class);
-            _jspx_th_form_005ferrors_005f1.setPageContext(_jspx_page_context);
-            _jspx_th_form_005ferrors_005f1.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_form_005fform_005f0);
-            // /pages/searchAd.jsp(83,2) name = path type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-            _jspx_th_form_005ferrors_005f1.setPath("city");
-            // /pages/searchAd.jsp(83,2) name = cssClass type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-            _jspx_th_form_005ferrors_005f1.setCssClass("validationErrorText");
-            int[] _jspx_push_body_count_form_005ferrors_005f1 = new int[] { 0 };
-            try {
-              int _jspx_eval_form_005ferrors_005f1 = _jspx_th_form_005ferrors_005f1.doStartTag();
-              if (_jspx_th_form_005ferrors_005f1.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-                return;
-              }
-            } catch (java.lang.Throwable _jspx_exception) {
-              while (_jspx_push_body_count_form_005ferrors_005f1[0]-- > 0)
-                out = _jspx_page_context.popBody();
-              _jspx_th_form_005ferrors_005f1.doCatch(_jspx_exception);
-            } finally {
-              _jspx_th_form_005ferrors_005f1.doFinally();
-              _005fjspx_005ftagPool_005fform_005ferrors_0026_005fpath_005fcssClass_005fnobody.reuse(_jspx_th_form_005ferrors_005f1);
-            }
-            out.write("\r\n");
-            out.write("\t\t\r\n");
-            out.write("\r\n");
-            out.write("\t\t<label for=\"radius\">Within radius of (max.):</label>\r\n");
-            out.write("\t\t");
-            if (_jspx_meth_form_005finput_005f1(_jspx_th_form_005fform_005f0, _jspx_page_context, _jspx_push_body_count_form_005fform_005f0))
-              return;
-            out.write("\r\n");
-            out.write("\t\tkm\r\n");
-            out.write("\t\t");
-            //  form:errors
-            org.springframework.web.servlet.tags.form.ErrorsTag _jspx_th_form_005ferrors_005f2 = (org.springframework.web.servlet.tags.form.ErrorsTag) _005fjspx_005ftagPool_005fform_005ferrors_0026_005fpath_005fcssClass_005fnobody.get(org.springframework.web.servlet.tags.form.ErrorsTag.class);
-            _jspx_th_form_005ferrors_005f2.setPageContext(_jspx_page_context);
-            _jspx_th_form_005ferrors_005f2.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_form_005fform_005f0);
-            // /pages/searchAd.jsp(90,2) name = path type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-            _jspx_th_form_005ferrors_005f2.setPath("radius");
-            // /pages/searchAd.jsp(90,2) name = cssClass type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-            _jspx_th_form_005ferrors_005f2.setCssClass("validationErrorText");
-            int[] _jspx_push_body_count_form_005ferrors_005f2 = new int[] { 0 };
-            try {
-              int _jspx_eval_form_005ferrors_005f2 = _jspx_th_form_005ferrors_005f2.doStartTag();
-              if (_jspx_th_form_005ferrors_005f2.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-                return;
-              }
-            } catch (java.lang.Throwable _jspx_exception) {
-              while (_jspx_push_body_count_form_005ferrors_005f2[0]-- > 0)
-                out = _jspx_page_context.popBody();
-              _jspx_th_form_005ferrors_005f2.doCatch(_jspx_exception);
-            } finally {
-              _jspx_th_form_005ferrors_005f2.doFinally();
-              _005fjspx_005ftagPool_005fform_005ferrors_0026_005fpath_005fcssClass_005fnobody.reuse(_jspx_th_form_005ferrors_005f2);
-            }
-            out.write("\r\n");
-            out.write("\t\t<br /> <label for=\"prize\">Price (max.):</label>\r\n");
-            out.write("\t\t");
-            if (_jspx_meth_form_005finput_005f2(_jspx_th_form_005fform_005f0, _jspx_page_context, _jspx_push_body_count_form_005fform_005f0))
-              return;
-            out.write("\r\n");
-            out.write("\t\tCHF\r\n");
-            out.write("\t\t");
-            //  form:errors
-            org.springframework.web.servlet.tags.form.ErrorsTag _jspx_th_form_005ferrors_005f3 = (org.springframework.web.servlet.tags.form.ErrorsTag) _005fjspx_005ftagPool_005fform_005ferrors_0026_005fpath_005fcssClass_005fnobody.get(org.springframework.web.servlet.tags.form.ErrorsTag.class);
-            _jspx_th_form_005ferrors_005f3.setPageContext(_jspx_page_context);
-            _jspx_th_form_005ferrors_005f3.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_form_005fform_005f0);
-            // /pages/searchAd.jsp(95,2) name = path type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-            _jspx_th_form_005ferrors_005f3.setPath("prize");
-            // /pages/searchAd.jsp(95,2) name = cssClass type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-            _jspx_th_form_005ferrors_005f3.setCssClass("validationErrorText");
-            int[] _jspx_push_body_count_form_005ferrors_005f3 = new int[] { 0 };
-            try {
-              int _jspx_eval_form_005ferrors_005f3 = _jspx_th_form_005ferrors_005f3.doStartTag();
-              if (_jspx_th_form_005ferrors_005f3.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-                return;
-              }
-            } catch (java.lang.Throwable _jspx_exception) {
-              while (_jspx_push_body_count_form_005ferrors_005f3[0]-- > 0)
-                out = _jspx_page_context.popBody();
-              _jspx_th_form_005ferrors_005f3.doCatch(_jspx_exception);
-            } finally {
-              _jspx_th_form_005ferrors_005f3.doFinally();
-              _005fjspx_005ftagPool_005fform_005ferrors_0026_005fpath_005fcssClass_005fnobody.reuse(_jspx_th_form_005ferrors_005f3);
-            }
-            out.write("\r\n");
-            out.write("\t\t<br />\r\n");
-            out.write("\r\n");
-            out.write("\t\t<button type=\"submit\" tabindex=\"7\" onClick=\"validateType(this.form)\">Search</button>\r\n");
-            out.write("\t\t<button type=\"reset\" tabindex=\"8\">Cancel</button>\r\n");
-            out.write("\t</fieldset>\r\n");
-            out.write("\r\n");
-            int evalDoAfterBody = _jspx_th_form_005fform_005f0.doAfterBody();
-            if (evalDoAfterBody != javax.servlet.jsp.tagext.BodyTag.EVAL_BODY_AGAIN)
-              break;
-          } while (true);
-        }
-        if (_jspx_th_form_005fform_005f0.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-          return;
-        }
-      } catch (java.lang.Throwable _jspx_exception) {
-        while (_jspx_push_body_count_form_005fform_005f0[0]-- > 0)
-          out = _jspx_page_context.popBody();
-        _jspx_th_form_005fform_005f0.doCatch(_jspx_exception);
-      } finally {
-        _jspx_th_form_005fform_005f0.doFinally();
-        _005fjspx_005ftagPool_005fform_005fform_0026_005fmodelAttribute_005fmethod_005fid_005fautocomplete_005faction.reuse(_jspx_th_form_005fform_005f0);
-      }
-      out.write("\r\n");
-      out.write("\r\n");
-      if (_jspx_meth_c_005fimport_005f2(_jspx_page_context))
-        return;
-    } catch (java.lang.Throwable t) {
-      if (!(t instanceof javax.servlet.jsp.SkipPageException)){
-        out = _jspx_out;
-        if (out != null && out.getBufferSize() != 0)
-          try {
-            if (response.isCommitted()) {
-              out.flush();
-            } else {
-              out.clearBuffer();
-            }
-          } catch (java.io.IOException e) {}
-        if (_jspx_page_context != null) _jspx_page_context.handlePageException(t);
-        else throw new ServletException(t);
-      }
-    } finally {
-      _jspxFactory.releasePageContext(_jspx_page_context);
-    }
-  }
-
-  private boolean _jspx_meth_c_005fimport_005f0(javax.servlet.jsp.PageContext _jspx_page_context)
-          throws java.lang.Throwable {
-    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
-    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
-    //  c:import
-    org.apache.taglibs.standard.tag.rt.core.ImportTag _jspx_th_c_005fimport_005f0 = (org.apache.taglibs.standard.tag.rt.core.ImportTag) _005fjspx_005ftagPool_005fc_005fimport_0026_005furl_005fnobody.get(org.apache.taglibs.standard.tag.rt.core.ImportTag.class);
-    _jspx_th_c_005fimport_005f0.setPageContext(_jspx_page_context);
-    _jspx_th_c_005fimport_005f0.setParent(null);
-    // /pages/searchAd.jsp(7,0) name = url type = null reqTime = true required = true fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_c_005fimport_005f0.setUrl("template/header.jsp");
-    int[] _jspx_push_body_count_c_005fimport_005f0 = new int[] { 0 };
-    try {
-      int _jspx_eval_c_005fimport_005f0 = _jspx_th_c_005fimport_005f0.doStartTag();
-      if (_jspx_th_c_005fimport_005f0.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-        return true;
-      }
-    } catch (java.lang.Throwable _jspx_exception) {
-      while (_jspx_push_body_count_c_005fimport_005f0[0]-- > 0)
-        out = _jspx_page_context.popBody();
-      _jspx_th_c_005fimport_005f0.doCatch(_jspx_exception);
-    } finally {
-      _jspx_th_c_005fimport_005f0.doFinally();
-      _005fjspx_005ftagPool_005fc_005fimport_0026_005furl_005fnobody.reuse(_jspx_th_c_005fimport_005f0);
-    }
-    return false;
-  }
-
-  private boolean _jspx_meth_c_005fimport_005f1(javax.servlet.jsp.PageContext _jspx_page_context)
-          throws java.lang.Throwable {
-    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
-    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
-    //  c:import
-    org.apache.taglibs.standard.tag.rt.core.ImportTag _jspx_th_c_005fimport_005f1 = (org.apache.taglibs.standard.tag.rt.core.ImportTag) _005fjspx_005ftagPool_005fc_005fimport_0026_005furl_005fnobody.get(org.apache.taglibs.standard.tag.rt.core.ImportTag.class);
-    _jspx_th_c_005fimport_005f1.setPageContext(_jspx_page_context);
-    _jspx_th_c_005fimport_005f1.setParent(null);
-    // /pages/searchAd.jsp(17,12) name = url type = null reqTime = true required = true fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_c_005fimport_005f1.setUrl("getzipcodes.jsp");
-    int[] _jspx_push_body_count_c_005fimport_005f1 = new int[] { 0 };
-    try {
-      int _jspx_eval_c_005fimport_005f1 = _jspx_th_c_005fimport_005f1.doStartTag();
-      if (_jspx_th_c_005fimport_005f1.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-        return true;
-      }
-    } catch (java.lang.Throwable _jspx_exception) {
-      while (_jspx_push_body_count_c_005fimport_005f1[0]-- > 0)
-        out = _jspx_page_context.popBody();
-      _jspx_th_c_005fimport_005f1.doCatch(_jspx_exception);
-    } finally {
-      _jspx_th_c_005fimport_005f1.doFinally();
-      _005fjspx_005ftagPool_005fc_005fimport_0026_005furl_005fnobody.reuse(_jspx_th_c_005fimport_005f1);
-    }
-    return false;
-  }
-
-  private boolean _jspx_meth_form_005fcheckbox_005f0(javax.servlet.jsp.tagext.JspTag _jspx_th_form_005fform_005f0, javax.servlet.jsp.PageContext _jspx_page_context, int[] _jspx_push_body_count_form_005fform_005f0)
-          throws java.lang.Throwable {
-    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
-    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
-    //  form:checkbox
-    org.springframework.web.servlet.tags.form.CheckboxTag _jspx_th_form_005fcheckbox_005f0 = (org.springframework.web.servlet.tags.form.CheckboxTag) _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fpath_005fname_005fid_005fnobody.get(org.springframework.web.servlet.tags.form.CheckboxTag.class);
-    _jspx_th_form_005fcheckbox_005f0.setPageContext(_jspx_page_context);
-    _jspx_th_form_005fcheckbox_005f0.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_form_005fform_005f0);
-    // /pages/searchAd.jsp(69,2) null
-    _jspx_th_form_005fcheckbox_005f0.setDynamicAttribute(null, "name", "room");
-    // /pages/searchAd.jsp(69,2) name = id type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005fcheckbox_005f0.setId("room");
-    // /pages/searchAd.jsp(69,2) name = path type = null reqTime = true required = true fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005fcheckbox_005f0.setPath("roomHelper");
-    int[] _jspx_push_body_count_form_005fcheckbox_005f0 = new int[] { 0 };
-    try {
-      int _jspx_eval_form_005fcheckbox_005f0 = _jspx_th_form_005fcheckbox_005f0.doStartTag();
-      if (_jspx_th_form_005fcheckbox_005f0.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-        return true;
-      }
-    } catch (java.lang.Throwable _jspx_exception) {
-      while (_jspx_push_body_count_form_005fcheckbox_005f0[0]-- > 0)
-        out = _jspx_page_context.popBody();
-      _jspx_th_form_005fcheckbox_005f0.doCatch(_jspx_exception);
-    } finally {
-      _jspx_th_form_005fcheckbox_005f0.doFinally();
-      _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fpath_005fname_005fid_005fnobody.reuse(_jspx_th_form_005fcheckbox_005f0);
-    }
-    return false;
-  }
-
-  private boolean _jspx_meth_form_005fcheckbox_005f1(javax.servlet.jsp.tagext.JspTag _jspx_th_form_005fform_005f0, javax.servlet.jsp.PageContext _jspx_page_context, int[] _jspx_push_body_count_form_005fform_005f0)
-          throws java.lang.Throwable {
-    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
-    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
-    //  form:checkbox
-    org.springframework.web.servlet.tags.form.CheckboxTag _jspx_th_form_005fcheckbox_005f1 = (org.springframework.web.servlet.tags.form.CheckboxTag) _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fpath_005fname_005fid_005fnobody.get(org.springframework.web.servlet.tags.form.CheckboxTag.class);
-    _jspx_th_form_005fcheckbox_005f1.setPageContext(_jspx_page_context);
-    _jspx_th_form_005fcheckbox_005f1.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_form_005fform_005f0);
-    // /pages/searchAd.jsp(70,2) null
-    _jspx_th_form_005fcheckbox_005f1.setDynamicAttribute(null, "name", "studio");
-    // /pages/searchAd.jsp(70,2) name = id type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005fcheckbox_005f1.setId("studio");
-    // /pages/searchAd.jsp(70,2) name = path type = null reqTime = true required = true fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005fcheckbox_005f1.setPath("studioHelper");
-    int[] _jspx_push_body_count_form_005fcheckbox_005f1 = new int[] { 0 };
-    try {
-      int _jspx_eval_form_005fcheckbox_005f1 = _jspx_th_form_005fcheckbox_005f1.doStartTag();
-      if (_jspx_th_form_005fcheckbox_005f1.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-        return true;
-      }
-    } catch (java.lang.Throwable _jspx_exception) {
-      while (_jspx_push_body_count_form_005fcheckbox_005f1[0]-- > 0)
-        out = _jspx_page_context.popBody();
-      _jspx_th_form_005fcheckbox_005f1.doCatch(_jspx_exception);
-    } finally {
-      _jspx_th_form_005fcheckbox_005f1.doFinally();
-      _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fpath_005fname_005fid_005fnobody.reuse(_jspx_th_form_005fcheckbox_005f1);
-    }
-    return false;
-  }
-
-  private boolean _jspx_meth_form_005fcheckbox_005f2(javax.servlet.jsp.tagext.JspTag _jspx_th_form_005fform_005f0, javax.servlet.jsp.PageContext _jspx_page_context, int[] _jspx_push_body_count_form_005fform_005f0)
-          throws java.lang.Throwable {
-    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
-    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
-    //  form:checkbox
-    org.springframework.web.servlet.tags.form.CheckboxTag _jspx_th_form_005fcheckbox_005f2 = (org.springframework.web.servlet.tags.form.CheckboxTag) _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fstyle_005fpath_005fname_005fid_005fnobody.get(org.springframework.web.servlet.tags.form.CheckboxTag.class);
-    _jspx_th_form_005fcheckbox_005f2.setPageContext(_jspx_page_context);
-    _jspx_th_form_005fcheckbox_005f2.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_form_005fform_005f0);
-    // /pages/searchAd.jsp(72,2) null
-    _jspx_th_form_005fcheckbox_005f2.setDynamicAttribute(null, "style", "display:none");
-    // /pages/searchAd.jsp(72,2) null
-    _jspx_th_form_005fcheckbox_005f2.setDynamicAttribute(null, "name", "neither");
-    // /pages/searchAd.jsp(72,2) name = id type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005fcheckbox_005f2.setId("neither");
-    // /pages/searchAd.jsp(72,2) name = path type = null reqTime = true required = true fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005fcheckbox_005f2.setPath("noRoomNoStudio");
-    int[] _jspx_push_body_count_form_005fcheckbox_005f2 = new int[] { 0 };
-    try {
-      int _jspx_eval_form_005fcheckbox_005f2 = _jspx_th_form_005fcheckbox_005f2.doStartTag();
-      if (_jspx_th_form_005fcheckbox_005f2.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-        return true;
-      }
-    } catch (java.lang.Throwable _jspx_exception) {
-      while (_jspx_push_body_count_form_005fcheckbox_005f2[0]-- > 0)
-        out = _jspx_page_context.popBody();
-      _jspx_th_form_005fcheckbox_005f2.doCatch(_jspx_exception);
-    } finally {
-      _jspx_th_form_005fcheckbox_005f2.doFinally();
-      _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fstyle_005fpath_005fname_005fid_005fnobody.reuse(_jspx_th_form_005fcheckbox_005f2);
-    }
-    return false;
-  }
-
-  private boolean _jspx_meth_form_005fcheckbox_005f3(javax.servlet.jsp.tagext.JspTag _jspx_th_form_005fform_005f0, javax.servlet.jsp.PageContext _jspx_page_context, int[] _jspx_push_body_count_form_005fform_005f0)
-          throws java.lang.Throwable {
-    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
-    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
-    //  form:checkbox
-    org.springframework.web.servlet.tags.form.CheckboxTag _jspx_th_form_005fcheckbox_005f3 = (org.springframework.web.servlet.tags.form.CheckboxTag) _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fstyle_005fpath_005fname_005fid_005fnobody.get(org.springframework.web.servlet.tags.form.CheckboxTag.class);
-    _jspx_th_form_005fcheckbox_005f3.setPageContext(_jspx_page_context);
-    _jspx_th_form_005fcheckbox_005f3.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_form_005fform_005f0);
-    // /pages/searchAd.jsp(73,2) null
-    _jspx_th_form_005fcheckbox_005f3.setDynamicAttribute(null, "style", "display:none");
-    // /pages/searchAd.jsp(73,2) null
-    _jspx_th_form_005fcheckbox_005f3.setDynamicAttribute(null, "name", "both");
-    // /pages/searchAd.jsp(73,2) name = id type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005fcheckbox_005f3.setId("both");
-    // /pages/searchAd.jsp(73,2) name = path type = null reqTime = true required = true fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005fcheckbox_005f3.setPath("bothRoomAndStudio");
-    int[] _jspx_push_body_count_form_005fcheckbox_005f3 = new int[] { 0 };
-    try {
-      int _jspx_eval_form_005fcheckbox_005f3 = _jspx_th_form_005fcheckbox_005f3.doStartTag();
-      if (_jspx_th_form_005fcheckbox_005f3.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-        return true;
-      }
-    } catch (java.lang.Throwable _jspx_exception) {
-      while (_jspx_push_body_count_form_005fcheckbox_005f3[0]-- > 0)
-        out = _jspx_page_context.popBody();
-      _jspx_th_form_005fcheckbox_005f3.doCatch(_jspx_exception);
-    } finally {
-      _jspx_th_form_005fcheckbox_005f3.doFinally();
-      _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fstyle_005fpath_005fname_005fid_005fnobody.reuse(_jspx_th_form_005fcheckbox_005f3);
-    }
-    return false;
-  }
-
-  private boolean _jspx_meth_form_005fcheckbox_005f4(javax.servlet.jsp.tagext.JspTag _jspx_th_form_005fform_005f0, javax.servlet.jsp.PageContext _jspx_page_context, int[] _jspx_push_body_count_form_005fform_005f0)
-          throws java.lang.Throwable {
-    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
-    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
-    //  form:checkbox
-    org.springframework.web.servlet.tags.form.CheckboxTag _jspx_th_form_005fcheckbox_005f4 = (org.springframework.web.servlet.tags.form.CheckboxTag) _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fstyle_005fpath_005fname_005fid_005fnobody.get(org.springframework.web.servlet.tags.form.CheckboxTag.class);
-    _jspx_th_form_005fcheckbox_005f4.setPageContext(_jspx_page_context);
-    _jspx_th_form_005fcheckbox_005f4.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_form_005fform_005f0);
-    // /pages/searchAd.jsp(74,2) null
-    _jspx_th_form_005fcheckbox_005f4.setDynamicAttribute(null, "style", "display:none");
-    // /pages/searchAd.jsp(74,2) null
-    _jspx_th_form_005fcheckbox_005f4.setDynamicAttribute(null, "name", "type");
-    // /pages/searchAd.jsp(74,2) name = id type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005fcheckbox_005f4.setId("type");
-    // /pages/searchAd.jsp(74,2) name = path type = null reqTime = true required = true fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005fcheckbox_005f4.setPath("studio");
-    int[] _jspx_push_body_count_form_005fcheckbox_005f4 = new int[] { 0 };
-    try {
-      int _jspx_eval_form_005fcheckbox_005f4 = _jspx_th_form_005fcheckbox_005f4.doStartTag();
-      if (_jspx_th_form_005fcheckbox_005f4.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-        return true;
-      }
-    } catch (java.lang.Throwable _jspx_exception) {
-      while (_jspx_push_body_count_form_005fcheckbox_005f4[0]-- > 0)
-        out = _jspx_page_context.popBody();
-      _jspx_th_form_005fcheckbox_005f4.doCatch(_jspx_exception);
-    } finally {
-      _jspx_th_form_005fcheckbox_005f4.doFinally();
-      _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fstyle_005fpath_005fname_005fid_005fnobody.reuse(_jspx_th_form_005fcheckbox_005f4);
-    }
-    return false;
-  }
-
-  private boolean _jspx_meth_form_005fcheckbox_005f5(javax.servlet.jsp.tagext.JspTag _jspx_th_form_005fform_005f0, javax.servlet.jsp.PageContext _jspx_page_context, int[] _jspx_push_body_count_form_005fform_005f0)
-          throws java.lang.Throwable {
-    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
-    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
-    //  form:checkbox
-    org.springframework.web.servlet.tags.form.CheckboxTag _jspx_th_form_005fcheckbox_005f5 = (org.springframework.web.servlet.tags.form.CheckboxTag) _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fstyle_005fpath_005fname_005fid_005fnobody.get(org.springframework.web.servlet.tags.form.CheckboxTag.class);
-    _jspx_th_form_005fcheckbox_005f5.setPageContext(_jspx_page_context);
-    _jspx_th_form_005fcheckbox_005f5.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_form_005fform_005f0);
-    // /pages/searchAd.jsp(75,2) null
-    _jspx_th_form_005fcheckbox_005f5.setDynamicAttribute(null, "style", "display:none");
-    // /pages/searchAd.jsp(75,2) null
-    _jspx_th_form_005fcheckbox_005f5.setDynamicAttribute(null, "name", "filtered");
-    // /pages/searchAd.jsp(75,2) name = id type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005fcheckbox_005f5.setId("filtered");
-    // /pages/searchAd.jsp(75,2) name = path type = null reqTime = true required = true fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005fcheckbox_005f5.setPath("filtered");
-    int[] _jspx_push_body_count_form_005fcheckbox_005f5 = new int[] { 0 };
-    try {
-      int _jspx_eval_form_005fcheckbox_005f5 = _jspx_th_form_005fcheckbox_005f5.doStartTag();
-      if (_jspx_th_form_005fcheckbox_005f5.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-        return true;
-      }
-    } catch (java.lang.Throwable _jspx_exception) {
-      while (_jspx_push_body_count_form_005fcheckbox_005f5[0]-- > 0)
-        out = _jspx_page_context.popBody();
-      _jspx_th_form_005fcheckbox_005f5.doCatch(_jspx_exception);
-    } finally {
-      _jspx_th_form_005fcheckbox_005f5.doFinally();
-      _005fjspx_005ftagPool_005fform_005fcheckbox_0026_005fstyle_005fpath_005fname_005fid_005fnobody.reuse(_jspx_th_form_005fcheckbox_005f5);
-    }
-    return false;
-  }
-
-  private boolean _jspx_meth_form_005finput_005f0(javax.servlet.jsp.tagext.JspTag _jspx_th_form_005fform_005f0, javax.servlet.jsp.PageContext _jspx_page_context, int[] _jspx_push_body_count_form_005fform_005f0)
-          throws java.lang.Throwable {
-    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
-    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
-    //  form:input
-    org.springframework.web.servlet.tags.form.InputTag _jspx_th_form_005finput_005f0 = (org.springframework.web.servlet.tags.form.InputTag) _005fjspx_005ftagPool_005fform_005finput_0026_005ftype_005ftabindex_005fplaceholder_005fpath_005fname_005fid_005fnobody.get(org.springframework.web.servlet.tags.form.InputTag.class);
-    _jspx_th_form_005finput_005f0.setPageContext(_jspx_page_context);
-    _jspx_th_form_005finput_005f0.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_form_005fform_005f0);
-    // /pages/searchAd.jsp(81,2) null
-    _jspx_th_form_005finput_005f0.setDynamicAttribute(null, "type", "text");
-    // /pages/searchAd.jsp(81,2) null
-    _jspx_th_form_005finput_005f0.setDynamicAttribute(null, "name", "city");
-    // /pages/searchAd.jsp(81,2) name = id type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005finput_005f0.setId("city");
-    // /pages/searchAd.jsp(81,2) name = path type = null reqTime = true required = true fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005finput_005f0.setPath("city");
-    // /pages/searchAd.jsp(81,2) null
-    _jspx_th_form_005finput_005f0.setDynamicAttribute(null, "placeholder", "e.g. Bern");
-    // /pages/searchAd.jsp(81,2) name = tabindex type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005finput_005f0.setTabindex("3");
-    int[] _jspx_push_body_count_form_005finput_005f0 = new int[] { 0 };
-    try {
-      int _jspx_eval_form_005finput_005f0 = _jspx_th_form_005finput_005f0.doStartTag();
-      if (_jspx_th_form_005finput_005f0.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-        return true;
-      }
-    } catch (java.lang.Throwable _jspx_exception) {
-      while (_jspx_push_body_count_form_005finput_005f0[0]-- > 0)
-        out = _jspx_page_context.popBody();
-      _jspx_th_form_005finput_005f0.doCatch(_jspx_exception);
-    } finally {
-      _jspx_th_form_005finput_005f0.doFinally();
-      _005fjspx_005ftagPool_005fform_005finput_0026_005ftype_005ftabindex_005fplaceholder_005fpath_005fname_005fid_005fnobody.reuse(_jspx_th_form_005finput_005f0);
-    }
-    return false;
-  }
-
-  private boolean _jspx_meth_form_005finput_005f1(javax.servlet.jsp.tagext.JspTag _jspx_th_form_005fform_005f0, javax.servlet.jsp.PageContext _jspx_page_context, int[] _jspx_push_body_count_form_005fform_005f0)
-          throws java.lang.Throwable {
-    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
-    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
-    //  form:input
-    org.springframework.web.servlet.tags.form.InputTag _jspx_th_form_005finput_005f1 = (org.springframework.web.servlet.tags.form.InputTag) _005fjspx_005ftagPool_005fform_005finput_0026_005ftype_005fstep_005fplaceholder_005fpath_005fid_005fnobody.get(org.springframework.web.servlet.tags.form.InputTag.class);
-    _jspx_th_form_005finput_005f1.setPageContext(_jspx_page_context);
-    _jspx_th_form_005finput_005f1.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_form_005fform_005f0);
-    // /pages/searchAd.jsp(87,2) name = id type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005finput_005f1.setId("radiusInput");
-    // /pages/searchAd.jsp(87,2) null
-    _jspx_th_form_005finput_005f1.setDynamicAttribute(null, "type", "number");
-    // /pages/searchAd.jsp(87,2) name = path type = null reqTime = true required = true fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005finput_005f1.setPath("radius");
-    // /pages/searchAd.jsp(87,2) null
-    _jspx_th_form_005finput_005f1.setDynamicAttribute(null, "placeholder", "e.g. 5");
-    // /pages/searchAd.jsp(87,2) null
-    _jspx_th_form_005finput_005f1.setDynamicAttribute(null, "step", "5");
-    int[] _jspx_push_body_count_form_005finput_005f1 = new int[] { 0 };
-    try {
-      int _jspx_eval_form_005finput_005f1 = _jspx_th_form_005finput_005f1.doStartTag();
-      if (_jspx_th_form_005finput_005f1.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-        return true;
-      }
-    } catch (java.lang.Throwable _jspx_exception) {
-      while (_jspx_push_body_count_form_005finput_005f1[0]-- > 0)
-        out = _jspx_page_context.popBody();
-      _jspx_th_form_005finput_005f1.doCatch(_jspx_exception);
-    } finally {
-      _jspx_th_form_005finput_005f1.doFinally();
-      _005fjspx_005ftagPool_005fform_005finput_0026_005ftype_005fstep_005fplaceholder_005fpath_005fid_005fnobody.reuse(_jspx_th_form_005finput_005f1);
-    }
-    return false;
-  }
-
-  private boolean _jspx_meth_form_005finput_005f2(javax.servlet.jsp.tagext.JspTag _jspx_th_form_005fform_005f0, javax.servlet.jsp.PageContext _jspx_page_context, int[] _jspx_push_body_count_form_005fform_005f0)
-          throws java.lang.Throwable {
-    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
-    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
-    //  form:input
-    org.springframework.web.servlet.tags.form.InputTag _jspx_th_form_005finput_005f2 = (org.springframework.web.servlet.tags.form.InputTag) _005fjspx_005ftagPool_005fform_005finput_0026_005ftype_005fstep_005fplaceholder_005fpath_005fid_005fnobody.get(org.springframework.web.servlet.tags.form.InputTag.class);
-    _jspx_th_form_005finput_005f2.setPageContext(_jspx_page_context);
-    _jspx_th_form_005finput_005f2.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_form_005fform_005f0);
-    // /pages/searchAd.jsp(92,2) name = id type = null reqTime = true required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005finput_005f2.setId("prizeInput");
-    // /pages/searchAd.jsp(92,2) null
-    _jspx_th_form_005finput_005f2.setDynamicAttribute(null, "type", "number");
-    // /pages/searchAd.jsp(92,2) name = path type = null reqTime = true required = true fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_form_005finput_005f2.setPath("prize");
-    // /pages/searchAd.jsp(92,2) null
-    _jspx_th_form_005finput_005f2.setDynamicAttribute(null, "placeholder", "e.g. 5");
-    // /pages/searchAd.jsp(92,2) null
-    _jspx_th_form_005finput_005f2.setDynamicAttribute(null, "step", "50");
-    int[] _jspx_push_body_count_form_005finput_005f2 = new int[] { 0 };
-    try {
-      int _jspx_eval_form_005finput_005f2 = _jspx_th_form_005finput_005f2.doStartTag();
-      if (_jspx_th_form_005finput_005f2.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-        return true;
-      }
-    } catch (java.lang.Throwable _jspx_exception) {
-      while (_jspx_push_body_count_form_005finput_005f2[0]-- > 0)
-        out = _jspx_page_context.popBody();
-      _jspx_th_form_005finput_005f2.doCatch(_jspx_exception);
-    } finally {
-      _jspx_th_form_005finput_005f2.doFinally();
-      _005fjspx_005ftagPool_005fform_005finput_0026_005ftype_005fstep_005fplaceholder_005fpath_005fid_005fnobody.reuse(_jspx_th_form_005finput_005f2);
-    }
-    return false;
-  }
-
-  private boolean _jspx_meth_c_005fimport_005f2(javax.servlet.jsp.PageContext _jspx_page_context)
-          throws java.lang.Throwable {
-    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
-    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
-    //  c:import
-    org.apache.taglibs.standard.tag.rt.core.ImportTag _jspx_th_c_005fimport_005f2 = (org.apache.taglibs.standard.tag.rt.core.ImportTag) _005fjspx_005ftagPool_005fc_005fimport_0026_005furl_005fnobody.get(org.apache.taglibs.standard.tag.rt.core.ImportTag.class);
-    _jspx_th_c_005fimport_005f2.setPageContext(_jspx_page_context);
-    _jspx_th_c_005fimport_005f2.setParent(null);
-    // /pages/searchAd.jsp(104,0) name = url type = null reqTime = true required = true fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
-    _jspx_th_c_005fimport_005f2.setUrl("template/footer.jsp");
-    int[] _jspx_push_body_count_c_005fimport_005f2 = new int[] { 0 };
-    try {
-      int _jspx_eval_c_005fimport_005f2 = _jspx_th_c_005fimport_005f2.doStartTag();
-      if (_jspx_th_c_005fimport_005f2.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
-        return true;
-      }
-    } catch (java.lang.Throwable _jspx_exception) {
-      while (_jspx_push_body_count_c_005fimport_005f2[0]-- > 0)
-        out = _jspx_page_context.popBody();
-      _jspx_th_c_005fimport_005f2.doCatch(_jspx_exception);
-    } finally {
-      _jspx_th_c_005fimport_005f2.doFinally();
-      _005fjspx_005ftagPool_005fc_005fimport_0026_005furl_005fnobody.reuse(_jspx_th_c_005fimport_005f2);
-    }
-    return false;
-  }
-}
